Split the AI settings page into separate Warp Agent and Agent Profiles pages
