+++ app/src/settings_view/agent_profiles_page.rs
@@ -1,0 +1,3123 @@
+//! The "Profiles" settings page, shown under the Agents umbrella.
+//!
+//! Owns the agent execution profiles and everything hanging off them: the
+//! per-profile permission dropdowns, the allow/denylist editors, the model
+//! pickers and the context-window control, plus the AI credit usage summary.
+
+use std::borrow::Cow;
+use std::cell::RefCell;
+use std::collections::HashMap;
+use std::path::{Path, PathBuf};
+
+use ::ai::api_keys::ApiKeyManager;
+use itertools::Itertools;
+use markdown_parser::{FormattedText, FormattedTextFragment, FormattedTextLine};
+use regex::Regex;
+use settings::{Setting, ToggleableSetting};
+use strum::IntoEnumIterator;
+use warp_core::features::FeatureFlag;
+use warp_errors::{report_error, report_if_error};
+use warpui::elements::{
+    ChildView, ConstrainedBox, Container, CrossAxisAlignment, Dismiss, Element, Flex,
+    FormattedTextElement, HighlightedHyperlink, HyperlinkLens, HyperlinkUrl, MainAxisAlignment,
+    MainAxisSize, MouseStateHandle, ParentElement, Shrinkable, Text,
+};
+use warpui::keymap::ContextPredicate;
+use warpui::ui_components::components::{Coords, UiComponent, UiComponentStyles};
+use warpui::ui_components::slider::SliderStateHandle;
+use warpui::ui_components::switch::SwitchStateHandle;
+use warpui::{
+    Action, AppContext, Entity, SingletonEntity, TypedActionView, View, ViewContext, ViewHandle,
+    WeakViewHandle, id,
+};
+
+use super::ai_shared::{
+    open_hyperlink, render_ai_setting_description, render_ai_setting_label,
+    render_ai_setting_toggle, should_show_mcp_servers, styles, update_editor_interaction_state,
+};
+use super::execution_profile_view::{ExecutionProfileView, ExecutionProfileViewEvent};
+use super::settings_page::{
+    CONTENT_FONT_SIZE, HEADER_PADDING, InputListItem, LocalOnlyIconState, MatchData, PageType,
+    SettingsPageMeta, SettingsPageViewHandle, SettingsWidget, ToggleState, build_sub_header,
+    render_body_item_label, render_body_item_label_with_icon, render_custom_size_header,
+    render_dropdown_item, render_dropdown_item_label, render_input_list, render_separator,
+    render_settings_info_banner,
+};
+use super::{SettingsAction, SettingsSection, ToggleSettingActionPair, flags};
+use crate::ai::blocklist::BlocklistAIPermissions;
+use crate::ai::execution_profiles::model_menu_items::available_model_menu_items;
+use crate::ai::execution_profiles::profiles::{
+    AIExecutionProfilesModel, AIExecutionProfilesModelEvent,
+};
+use crate::ai::execution_profiles::{
+    AIExecutionProfile, AIExecutionProfileAppExt, ActionPermission, ExecutionProfileId,
+    WriteToPtyPermission, long_context_pricing_warning_title,
+};
+use crate::ai::llms::{LLMContextWindow, LLMId, LLMPreferences, LLMPreferencesEvent};
+use crate::ai::mcp::TemplatableMCPServerManager;
+use crate::ai::paths::host_native_absolute_path;
+use crate::ai::{AIRequestUsageModel, AIRequestUsageModelEvent};
+use crate::appearance::Appearance;
+use crate::auth::AuthStateProvider;
+use crate::auth::auth_manager::{AuthManager, LoginGatedFeature};
+use crate::auth::auth_view_modal::AuthViewVariant;
+use crate::cloud_object::GenericStringObjectFormat::Json;
+use crate::cloud_object::model::persistence::{CloudModel, CloudModelEvent};
+use crate::cloud_object::{JsonObjectType, ObjectType};
+use crate::editor::{EditorView, Event as EditorEvent, SingleLineEditorOptions, TextOptions};
+use crate::server::telemetry::AutonomySettingToggleSource;
+use crate::settings::{
+    AISettings, AISettingsChangedEvent, AgentModeCodingPermissionsType,
+    AgentModeCommandExecutionDenylist, AgentModeCommandExecutionPredicate, CodeSettings,
+    CodebaseContextEnabled,
+};
+use crate::terminal::session_settings::{SessionSettings, SessionSettingsChangedEvent};
+use crate::ui_components::blended_colors;
+use crate::ui_components::icons::Icon;
+use crate::util::bindings;
+use crate::view_components::action_button::{ActionButton, ButtonSize, SecondaryTheme};
+use crate::view_components::dropdown::DropdownAction;
+use crate::view_components::{
+    Dropdown, DropdownItem, FilterableDropdown, SubmittableTextInput, SubmittableTextInputEvent,
+    WarningBoxConfig, render_warning_box,
+};
+use crate::workspaces::user_workspaces::UserWorkspacesEvent;
+use crate::{TelemetryEvent, UserWorkspaces, send_telemetry_from_ctx};
+
+const AI_SETTINGS_DROPDOWN_WIDTH: f32 = 250.;
+const AI_SETTINGS_DROPDOWN_MAX_HEIGHT: f32 = 250.;
+const CONTEXT_WINDOW_SLIDER_WIDTH: f32 = 220.;
+const CONTEXT_WINDOW_INPUT_BOX_WIDTH: f32 = 120.;
+
+pub struct AgentProfilesPageView {
+    page: PageType<Self>,
+    local_only_icon_tooltip_states: RefCell<HashMap<String, MouseStateHandle>>,
+
+    autonomy_dropdown_menu: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+    code_read_autonomy_dropdown_menu: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+
+    code_read_allowlist_editor: ViewHandle<SubmittableTextInput>,
+    code_read_allowlist_mouse_state_handles: Vec<MouseStateHandle>,
+
+    command_execution_allowlist_editor: ViewHandle<SubmittableTextInput>,
+    command_execution_allowlist_mouse_state_handles: Vec<MouseStateHandle>,
+    command_execution_denylist_editor: ViewHandle<SubmittableTextInput>,
+    command_execution_denylist_mouse_state_handles: Vec<MouseStateHandle>,
+
+    apply_code_diffs_dropdown_menu: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+    read_files_dropdown_menu: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+    execute_commands_dropdown_menu: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+    write_to_pty_autonomy_dropdown_menu: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+    mcp_permissions_dropdown_menu: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+
+    // Allowlisting directories (default profile)
+    directory_allowlist_mouse_state_handles: Vec<MouseStateHandle>,
+    directory_allowlist_editor: ViewHandle<SubmittableTextInput>,
+
+    // Allowlisting commands (default profile)
+    command_allowlist_mouse_state_handles: Vec<MouseStateHandle>,
+    command_allowlist_editor: ViewHandle<SubmittableTextInput>,
+
+    // Denylisting commands (default profile)
+    command_denylist_mouse_state_handles: Vec<MouseStateHandle>,
+    command_denylist_tooltip_mouse_state_handles: Vec<MouseStateHandle>,
+    command_denylist_editor: ViewHandle<SubmittableTextInput>,
+
+    mcp_allowlist_mouse_state_handles: Vec<MouseStateHandle>,
+    mcp_allowlist_dropdown: ViewHandle<FilterableDropdown<AgentProfilesPageAction>>,
+
+    mcp_denylist_mouse_state_handles: Vec<MouseStateHandle>,
+    mcp_denylist_dropdown: ViewHandle<FilterableDropdown<AgentProfilesPageAction>>,
+
+    base_model_dropdown: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+    coding_model_dropdown: ViewHandle<Dropdown<AgentProfilesPageAction>>,
+
+    context_window_slider_state: SliderStateHandle,
+    context_window_editor: ViewHandle<EditorView>,
+    last_synced_context_window_editor_value: Option<u32>,
+    dragged_context_window_value: Option<u32>,
+
+    // Profile views
+    profile_views: Vec<ViewHandle<ExecutionProfileView>>,
+    add_profile_button: ViewHandle<ActionButton>,
+}
+
+impl AgentProfilesPageView {
+    pub fn new(ctx: &mut ViewContext<Self>) -> Self {
+        let is_any_ai_enabled = AISettings::as_ref(ctx).is_any_ai_enabled(ctx);
+
+        let workspace = UserWorkspaces::handle(ctx);
+        let ai_autonomy_settings = workspace.as_ref(ctx).ai_autonomy_settings();
+        ctx.subscribe_to_model(&workspace, |me, workspace, event, ctx| {
+            if let UserWorkspacesEvent::TeamsChanged = event {
+                me.refresh_all_execution_profile_ui(ctx);
+                me.reset_execution_profile_mouse_state_handles(ctx);
+
+                let is_any_ai_enabled = AISettings::as_ref(ctx).is_any_ai_enabled(ctx);
+                let ai_autonomy_settings = workspace.as_ref(ctx).ai_autonomy_settings();
+
+                update_editor_interaction_state(
+                    me.command_denylist_editor.as_ref(ctx).editor().clone(),
+                    is_any_ai_enabled,
+                    ctx,
+                );
+
+                update_editor_interaction_state(
+                    me.command_allowlist_editor.as_ref(ctx).editor().clone(),
+                    is_any_ai_enabled
+                        && !ai_autonomy_settings.has_override_for_execute_commands_allowlist(),
+                    ctx,
+                );
+
+                update_editor_interaction_state(
+                    me.directory_allowlist_editor.as_ref(ctx).editor().clone(),
+                    is_any_ai_enabled
+                        && !ai_autonomy_settings.has_override_for_read_files_allowlist(),
+                    ctx,
+                );
+
+                ctx.notify();
+            }
+        });
+
+        let coding_model_dropdown = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown.set_menu_max_height(AI_SETTINGS_DROPDOWN_MAX_HEIGHT, ctx);
+            dropdown
+        });
+        Self::refresh_coding_model_menu(&coding_model_dropdown, ctx);
+
+        let base_model_dropdown = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown.set_menu_max_height(AI_SETTINGS_DROPDOWN_MAX_HEIGHT, ctx);
+
+            dropdown
+        });
+        Self::refresh_base_model_menu(&base_model_dropdown, ctx);
+
+        let initial_context_window_value = Self::initial_context_window_value(ctx);
+        let clamped_initial = Self::configurable_context_window(ctx)
+            .map(|cw| initial_context_window_value.clamp(cw.min, cw.max))
+            .unwrap_or(initial_context_window_value);
+        let context_window_slider_state = SliderStateHandle::default();
+
+        let context_window_editor = ctx.add_typed_action_view(|ctx| {
+            let options = SingleLineEditorOptions {
+                text: TextOptions {
+                    font_size_override: Some(Appearance::as_ref(ctx).ui_font_size()),
+                    ..Default::default()
+                },
+                ..Default::default()
+            };
+            let mut editor = EditorView::single_line(options, ctx);
+            editor.set_buffer_text(&clamped_initial.to_string(), ctx);
+            editor
+        });
+        ctx.subscribe_to_view(&context_window_editor, |me, _, event, ctx| {
+            me.handle_context_window_editor_event(event, ctx);
+        });
+        let last_synced_context_window_editor_value = Some(clamped_initial);
+
+        let autonomy_dropdown_menu = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown
+        });
+        Self::refresh_autonomy_dropdown_menu(&autonomy_dropdown_menu, ctx);
+
+        let code_read_autonomy_dropdown_menu = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown
+        });
+        Self::refresh_code_read_autonomy_dropdown_menu(&code_read_autonomy_dropdown_menu, ctx);
+
+        // While the data model supports arbitrary files in the allowlist,
+        // it's most intuitive to allowlist whole directories.
+        let code_read_allowlist_editor = ctx.add_typed_action_view(|ctx| {
+            let mut input = SubmittableTextInput::new(ctx).validate_on_submit(|s| {
+                let expanded = host_native_absolute_path(s, &None, &None);
+                Path::new(&expanded).is_dir()
+            });
+            input.set_placeholder_text("e.g. ~/code-repos/repo", ctx);
+            input
+        });
+        update_editor_interaction_state(
+            code_read_allowlist_editor.as_ref(ctx).editor().clone(),
+            is_any_ai_enabled,
+            ctx,
+        );
+
+        ctx.subscribe_to_view(&code_read_allowlist_editor, |_, _, event, ctx| {
+            if let SubmittableTextInputEvent::Submit(s) = event {
+                let expanded = host_native_absolute_path(s, &None, &None);
+                BlocklistAIPermissions::handle(ctx).update(ctx, |model, ctx| {
+                    report_if_error!(
+                        model.add_filepath_to_code_read_allowlist(PathBuf::from(expanded), ctx)
+                    );
+                });
+            }
+        });
+
+        let command_execution_allowlist_editor = ctx.add_typed_action_view(|ctx| {
+            let mut input =
+                SubmittableTextInput::new(ctx).validate_on_edit(|s| Regex::new(s).is_ok());
+            input.set_placeholder_text("e.g. ls .*", ctx);
+            input
+        });
+        update_editor_interaction_state(
+            command_execution_allowlist_editor
+                .as_ref(ctx)
+                .editor()
+                .clone(),
+            is_any_ai_enabled,
+            ctx,
+        );
+
+        ctx.subscribe_to_view(&command_execution_allowlist_editor, |_, _, event, ctx| {
+            if let SubmittableTextInputEvent::Submit(s) = event {
+                let predicate = match AgentModeCommandExecutionPredicate::new_regex(s) {
+                    Ok(regex) => regex,
+                    Err(e) => {
+                        log::warn!(
+                            "Failed to convert string to regex for cmd execution allowlist: {e}"
+                        );
+                        return;
+                    }
+                };
+                BlocklistAIPermissions::handle(ctx).update(ctx, |model, ctx| {
+                    report_if_error!(model.add_command_to_autoexecution_allowlist(predicate, ctx));
+                })
+            }
+        });
+
+        let command_execution_denylist_editor = ctx.add_typed_action_view(|ctx| {
+            let mut input =
+                SubmittableTextInput::new(ctx).validate_on_edit(|s| Regex::new(s).is_ok());
+            input.set_placeholder_text("e.g. rm .*", ctx);
+            input
+        });
+        update_editor_interaction_state(
+            command_execution_denylist_editor
+                .as_ref(ctx)
+                .editor()
+                .clone(),
+            is_any_ai_enabled,
+            ctx,
+        );
+
+        ctx.subscribe_to_view(&command_execution_denylist_editor, |_, _, event, ctx| {
+            if let SubmittableTextInputEvent::Submit(s) = event {
+                let predicate = match AgentModeCommandExecutionPredicate::new_regex(s) {
+                    Ok(regex) => regex,
+                    Err(e) => {
+                        log::warn!(
+                            "Failed to convert string to regex for cmd execution denylist: {e}"
+                        );
+                        return;
+                    }
+                };
+                BlocklistAIPermissions::handle(ctx).update(ctx, |model, ctx| {
+                    report_if_error!(model.add_command_to_autoexecution_denylist(predicate, ctx));
+                })
+            }
+        });
+
+        ctx.subscribe_to_model(&UserWorkspaces::handle(ctx), |me, _handle, _event, ctx| {
+            // Re-render if teams-related data changed that may affect which
+            // models are available.
+            Self::refresh_base_model_menu(&me.base_model_dropdown, ctx);
+            Self::refresh_coding_model_menu(&me.coding_model_dropdown, ctx);
+            ctx.notify();
+        });
+
+        ctx.subscribe_to_model(
+            &AIExecutionProfilesModel::handle(ctx),
+            |me, _, event, ctx| {
+                match event {
+                    AIExecutionProfilesModelEvent::ProfileCreated
+                    | AIExecutionProfilesModelEvent::ProfileDeleted => {
+                        me.refresh_profile_views(ctx);
+                    }
+                    AIExecutionProfilesModelEvent::ProfileUpdated(_) => {
+                        me.refresh_all_execution_profile_ui(ctx);
+                        me.reset_execution_profile_mouse_state_handles(ctx);
+                        me.sync_context_window_editor(ctx, false);
+                    }
+                    AIExecutionProfilesModelEvent::UpdatedActiveProfile { .. } => (),
+                }
+                ctx.notify();
+            },
+        );
+
+        let cloud_model = CloudModel::handle(ctx);
+        ctx.subscribe_to_model(&cloud_model, |me, _, event, ctx| {
+            let added_or_deleted_mcp_servers = matches!(
+                event,
+                CloudModelEvent::ObjectCreated { type_and_id } | CloudModelEvent::ObjectDeleted { type_and_id, .. }
+                if matches!(
+                    type_and_id.object_type(),
+                    ObjectType::GenericStringObject(Json(JsonObjectType::MCPServer))
+                )
+            );
+
+            if added_or_deleted_mcp_servers {
+                Self::refresh_mcp_allowlist_dropdown(&me.mcp_allowlist_dropdown, ctx);
+                Self::refresh_mcp_denylist_dropdown(&me.mcp_denylist_dropdown, ctx);
+                ctx.notify();
+            }
+        });
+
+        let templatable_manager = TemplatableMCPServerManager::handle(ctx);
+        ctx.subscribe_to_model(&templatable_manager, |me, _, _event, ctx| {
+            Self::refresh_mcp_allowlist_dropdown(&me.mcp_allowlist_dropdown, ctx);
+            Self::refresh_mcp_denylist_dropdown(&me.mcp_denylist_dropdown, ctx);
+            ctx.notify();
+        });
+
+        ctx.subscribe_to_model(
+            &LLMPreferences::handle(ctx),
+            |me, _, event, ctx| match event {
+                LLMPreferencesEvent::UpdatedAvailableLLMs => {
+                    Self::refresh_base_model_menu(&me.base_model_dropdown, ctx);
+                    Self::refresh_coding_model_menu(&me.coding_model_dropdown, ctx);
+                    me.sync_context_window_editor(ctx, false);
+                }
+                LLMPreferencesEvent::UpdatedActiveAgentModeLLM => {
+                    Self::refresh_base_model_menu(&me.base_model_dropdown, ctx);
+                    me.sync_context_window_editor(ctx, false);
+                }
+                LLMPreferencesEvent::UpdatedActiveCodingLLM => {
+                    Self::refresh_coding_model_menu(&me.coding_model_dropdown, ctx);
+                }
+            },
+        );
+
+        // Refresh model dropdowns when BYO API keys update so key icons reflect latest state.
+        ctx.subscribe_to_model(&ApiKeyManager::handle(ctx), |me, _model, _event, ctx| {
+            Self::refresh_base_model_menu(&me.base_model_dropdown, ctx);
+            Self::refresh_coding_model_menu(&me.coding_model_dropdown, ctx);
+            me.sync_context_window_editor(ctx, false);
+            ctx.notify();
+        });
+
+        ctx.subscribe_to_model(&AISettings::handle(ctx), |me, _, event, ctx| {
+            match event {
+                AISettingsChangedEvent::IsAnyAIEnabled { .. } => {
+                    let is_enabled = AISettings::as_ref(ctx).is_any_ai_enabled(ctx);
+                    let ai_autonomy_settings = UserWorkspaces::as_ref(ctx).ai_autonomy_settings();
+
+                    update_editor_interaction_state(
+                        me.command_execution_allowlist_editor
+                            .as_ref(ctx)
+                            .editor()
+                            .clone(),
+                        is_enabled,
+                        ctx,
+                    );
+                    update_editor_interaction_state(
+                        me.command_execution_denylist_editor
+                            .as_ref(ctx)
+                            .editor()
+                            .clone(),
+                        is_enabled,
+                        ctx,
+                    );
+                    update_editor_interaction_state(
+                        me.code_read_allowlist_editor.as_ref(ctx).editor().clone(),
+                        is_enabled,
+                        ctx,
+                    );
+
+                    update_editor_interaction_state(
+                        me.directory_allowlist_editor.as_ref(ctx).editor().clone(),
+                        is_enabled && !ai_autonomy_settings.has_override_for_read_files_allowlist(),
+                        ctx,
+                    );
+
+                    update_editor_interaction_state(
+                        me.command_denylist_editor.as_ref(ctx).editor().clone(),
+                        is_enabled,
+                        ctx,
+                    );
+
+                    update_editor_interaction_state(
+                        me.command_allowlist_editor.as_ref(ctx).editor().clone(),
+                        is_enabled
+                            && !ai_autonomy_settings.has_override_for_execute_commands_allowlist(),
+                        ctx,
+                    );
+
+                    Self::refresh_autonomy_dropdown_menu(&me.autonomy_dropdown_menu, ctx);
+
+                    me.refresh_all_execution_profile_ui(ctx);
+
+                    Self::refresh_code_read_autonomy_dropdown_menu(
+                        &me.code_read_autonomy_dropdown_menu,
+                        ctx,
+                    );
+                    Self::refresh_base_model_menu(&me.base_model_dropdown, ctx);
+                    Self::refresh_coding_model_menu(&me.coding_model_dropdown, ctx);
+                    Self::refresh_mcp_allowlist_dropdown(&me.mcp_allowlist_dropdown, ctx);
+                    Self::refresh_mcp_denylist_dropdown(&me.mcp_denylist_dropdown, ctx);
+                    me.sync_context_window_editor(ctx, true);
+                }
+                AISettingsChangedEvent::AgentModeExecuteReadonlyCommands { .. } => {
+                    Self::refresh_autonomy_dropdown_menu(&me.autonomy_dropdown_menu, ctx);
+                    Self::refresh_code_read_autonomy_dropdown_menu(
+                        &me.code_read_autonomy_dropdown_menu,
+                        ctx,
+                    );
+                }
+                AISettingsChangedEvent::AgentModeCodingPermissions { .. } => {
+                    Self::refresh_code_read_autonomy_dropdown_menu(
+                        &me.code_read_autonomy_dropdown_menu,
+                        ctx,
+                    );
+                }
+                AISettingsChangedEvent::AgentModeCommandExecutionAllowlist { .. } => {
+                    me.command_execution_allowlist_mouse_state_handles = AISettings::as_ref(ctx)
+                        .agent_mode_command_execution_allowlist
+                        .value()
+                        .iter()
+                        .map(|_| Default::default())
+                        .collect();
+                }
+                AISettingsChangedEvent::AgentModeCommandExecutionDenylist { .. } => {
+                    me.command_execution_denylist_mouse_state_handles = AISettings::as_ref(ctx)
+                        .agent_mode_command_execution_denylist
+                        .value()
+                        .iter()
+                        .map(|_| Default::default())
+                        .collect();
+                }
+                AISettingsChangedEvent::AgentModeCodingFileReadAllowlist { .. } => {
+                    me.code_read_allowlist_mouse_state_handles = AISettings::as_ref(ctx)
+                        .agent_mode_coding_file_read_allowlist
+                        .value()
+                        .iter()
+                        .map(|_| Default::default())
+                        .collect();
+                }
+                _ => (),
+            }
+            ctx.notify();
+        });
+
+        ctx.subscribe_to_model(&SessionSettings::handle(ctx), |_, _, event, ctx| {
+            if let SessionSettingsChangedEvent::ShowModelSelectorsInPrompt { .. } = event {
+                ctx.notify();
+            }
+        });
+
+        let current_permission =
+            BlocklistAIPermissions::as_ref(ctx).active_permissions_profile(ctx, None);
+
+        let apply_code_diffs_dropdown_menu = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+
+            dropdown.set_items(
+                vec![
+                    DropdownItem::new(
+                        "Agent decides",
+                        AgentProfilesPageAction::SetApplyCodeDiffs(ActionPermission::AgentDecides),
+                    ),
+                    DropdownItem::new(
+                        "Always allow",
+                        AgentProfilesPageAction::SetApplyCodeDiffs(ActionPermission::AlwaysAllow),
+                    ),
+                    DropdownItem::new(
+                        "Always ask",
+                        AgentProfilesPageAction::SetApplyCodeDiffs(ActionPermission::AlwaysAsk),
+                    ),
+                ],
+                ctx,
+            );
+            dropdown
+        });
+        Self::refresh_execution_profile_dropdown_menu(
+            &apply_code_diffs_dropdown_menu,
+            current_permission.apply_code_diffs,
+            !AISettings::as_ref(ctx).is_code_diffs_permissions_editable(ctx),
+            ctx,
+        );
+
+        let read_files_dropdown_menu = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown.set_items(
+                vec![
+                    DropdownItem::new(
+                        "Agent decides",
+                        AgentProfilesPageAction::SetReadFiles(ActionPermission::AgentDecides),
+                    ),
+                    DropdownItem::new(
+                        "Always allow",
+                        AgentProfilesPageAction::SetReadFiles(ActionPermission::AlwaysAllow),
+                    ),
+                    DropdownItem::new(
+                        "Always ask",
+                        AgentProfilesPageAction::SetReadFiles(ActionPermission::AlwaysAsk),
+                    ),
+                ],
+                ctx,
+            );
+            dropdown
+        });
+        Self::refresh_execution_profile_dropdown_menu(
+            &read_files_dropdown_menu,
+            current_permission.read_files,
+            !AISettings::as_ref(ctx).is_read_files_permissions_editable(ctx),
+            ctx,
+        );
+
+        let execute_commands_dropdown_menu = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown.set_items(
+                vec![
+                    DropdownItem::new(
+                        "Agent decides",
+                        AgentProfilesPageAction::SetExecuteCommands(ActionPermission::AgentDecides),
+                    ),
+                    DropdownItem::new(
+                        "Always allow",
+                        AgentProfilesPageAction::SetExecuteCommands(ActionPermission::AlwaysAllow),
+                    ),
+                    DropdownItem::new(
+                        "Always ask",
+                        AgentProfilesPageAction::SetExecuteCommands(ActionPermission::AlwaysAsk),
+                    ),
+                ],
+                ctx,
+            );
+            dropdown
+        });
+        Self::refresh_execution_profile_dropdown_menu(
+            &execute_commands_dropdown_menu,
+            current_permission.execute_commands,
+            !AISettings::as_ref(ctx).is_execute_commands_permissions_editable(ctx),
+            ctx,
+        );
+
+        let write_to_pty_autonomy_dropdown_menu = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown.set_items(
+                vec![
+                    DropdownItem::new(
+                        "Always allow",
+                        AgentProfilesPageAction::SetWriteToPty(WriteToPtyPermission::AlwaysAllow),
+                    ),
+                    DropdownItem::new(
+                        "Always ask",
+                        AgentProfilesPageAction::SetWriteToPty(WriteToPtyPermission::AlwaysAsk),
+                    ),
+                    DropdownItem::new(
+                        "Ask on first write",
+                        AgentProfilesPageAction::SetWriteToPty(
+                            WriteToPtyPermission::AskOnFirstWrite,
+                        ),
+                    ),
+                ],
+                ctx,
+            );
+            dropdown
+        });
+        Self::refresh_write_to_pty_dropdown_menu(
+            &write_to_pty_autonomy_dropdown_menu,
+            current_permission.write_to_pty,
+            !AISettings::as_ref(ctx).is_write_to_pty_permissions_editable(ctx),
+            ctx,
+        );
+
+        let mcp_permissions_dropdown_menu = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = Dropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown.set_items(
+                vec![
+                    DropdownItem::new(
+                        "Agent decides",
+                        AgentProfilesPageAction::SetMCPPermissions(ActionPermission::AgentDecides),
+                    ),
+                    DropdownItem::new(
+                        "Always allow",
+                        AgentProfilesPageAction::SetMCPPermissions(ActionPermission::AlwaysAllow),
+                    ),
+                    DropdownItem::new(
+                        "Always ask",
+                        AgentProfilesPageAction::SetMCPPermissions(ActionPermission::AlwaysAsk),
+                    ),
+                ],
+                ctx,
+            );
+            dropdown
+        });
+        Self::refresh_execution_profile_dropdown_menu(
+            &mcp_permissions_dropdown_menu,
+            current_permission.mcp_permissions,
+            !AISettings::as_ref(ctx).is_mcp_permission_editable(ctx),
+            ctx,
+        );
+
+        let mcp_allowlist_dropdown = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = FilterableDropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown.set_menu_header_to_static("Select MCP servers");
+            dropdown
+        });
+        Self::refresh_mcp_allowlist_dropdown(&mcp_allowlist_dropdown, ctx);
+        let mcp_allowlist_mouse_state_handles = BlocklistAIPermissions::as_ref(ctx)
+            .get_mcp_allowlist(ctx, None)
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        let mcp_denylist_dropdown = ctx.add_typed_action_view(|ctx| {
+            let mut dropdown = FilterableDropdown::new(ctx);
+            dropdown.set_top_bar_max_width(AI_SETTINGS_DROPDOWN_WIDTH);
+            dropdown.set_menu_width(AI_SETTINGS_DROPDOWN_WIDTH, ctx);
+            dropdown.set_menu_header_to_static("Select MCP servers");
+            dropdown
+        });
+        Self::refresh_mcp_denylist_dropdown(&mcp_denylist_dropdown, ctx);
+        let mcp_denylist_mouse_state_handles = BlocklistAIPermissions::as_ref(ctx)
+            .get_mcp_denylist(ctx, None)
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        let command_execution_allowlist_mouse_state_handles = AISettings::as_ref(ctx)
+            .agent_mode_command_execution_allowlist
+            .value()
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        let command_execution_denylist_mouse_state_handles = AISettings::as_ref(ctx)
+            .agent_mode_command_execution_denylist
+            .value()
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+        let code_read_allowlist_mouse_state_handles = AISettings::as_ref(ctx)
+            .agent_mode_coding_file_read_allowlist
+            .value()
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        let directory_allowlist_mouse_state_handles = current_permission
+            .directory_allowlist
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        let directory_allowlist_editor = ctx.add_typed_action_view(|ctx| {
+            let mut input = SubmittableTextInput::new(ctx).validate_on_submit(|s| {
+                let expanded = host_native_absolute_path(s, &None, &None);
+                Path::new(&expanded).is_dir()
+            });
+            input.set_placeholder_text("e.g. ~/code-repos/repo", ctx);
+            input
+        });
+
+        update_editor_interaction_state(
+            directory_allowlist_editor.as_ref(ctx).editor().clone(),
+            is_any_ai_enabled,
+            ctx,
+        );
+
+        ctx.subscribe_to_view(&directory_allowlist_editor, |_, _, event, ctx| {
+            if let SubmittableTextInputEvent::Submit(s) = event {
+                let expanded = host_native_absolute_path(s, &None, &None);
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+
+                    model.add_to_directory_allowlist(profile_id, &PathBuf::from(expanded), ctx);
+                });
+                ctx.notify();
+            }
+        });
+
+        let org_denylist = BlocklistAIPermissions::get_org_execute_commands_denylist(ctx);
+        let command_denylist_mouse_state_handles = current_permission
+            .command_denylist
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+        let command_denylist_tooltip_mouse_state_handles: Vec<MouseStateHandle> =
+            org_denylist.iter().map(|_| Default::default()).collect();
+
+        let command_denylist_editor = ctx.add_typed_action_view(|ctx| {
+            let mut input =
+                SubmittableTextInput::new(ctx).validate_on_edit(|s| Regex::new(s).is_ok());
+            input.set_placeholder_text("e.g. rm .*", ctx);
+            input
+        });
+        update_editor_interaction_state(
+            command_denylist_editor.as_ref(ctx).editor().clone(),
+            is_any_ai_enabled && !ai_autonomy_settings.has_override_for_execute_commands_denylist(),
+            ctx,
+        );
+
+        ctx.subscribe_to_view(&command_denylist_editor, |_, _, event, ctx| {
+            if let SubmittableTextInputEvent::Submit(s) = event {
+                let predicate = match AgentModeCommandExecutionPredicate::new_regex(s) {
+                    Ok(regex) => regex,
+                    Err(e) => {
+                        log::warn!(
+                            "Failed to convert string to regex for cmd execution denylist: {e}"
+                        );
+                        return;
+                    }
+                };
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+                    model.add_to_command_denylist(profile_id, &predicate, ctx);
+                });
+                ctx.notify();
+            }
+        });
+
+        let command_allowlist_mouse_state_handles = current_permission
+            .command_allowlist
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        let command_allowlist_editor = ctx.add_typed_action_view(|ctx| {
+            let mut input =
+                SubmittableTextInput::new(ctx).validate_on_edit(|s| Regex::new(s).is_ok());
+            input.set_placeholder_text("e.g. ls .*", ctx);
+            input
+        });
+        update_editor_interaction_state(
+            command_allowlist_editor.as_ref(ctx).editor().clone(),
+            is_any_ai_enabled
+                && !ai_autonomy_settings.has_override_for_execute_commands_allowlist(),
+            ctx,
+        );
+
+        ctx.subscribe_to_view(&command_allowlist_editor, |_, _, event, ctx| {
+            if let SubmittableTextInputEvent::Submit(s) = event {
+                let predicate = match AgentModeCommandExecutionPredicate::new_regex(s) {
+                    Ok(regex) => regex,
+                    Err(e) => {
+                        log::warn!(
+                            "Failed to convert string to regex for cmd execution allowlist: {e}"
+                        );
+                        return;
+                    }
+                };
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+                    model.add_to_command_allowlist(profile_id, &predicate, ctx);
+                });
+                ctx.notify();
+            }
+        });
+
+        let ai_request_model = AIRequestUsageModel::handle(ctx);
+        ctx.subscribe_to_model(&ai_request_model, |me, _, event, ctx| {
+            match event {
+                AIRequestUsageModelEvent::RequestUsageUpdated => ctx.notify(),
+                AIRequestUsageModelEvent::CreditAvailabilityUpdated => ctx.notify(),
+                AIRequestUsageModelEvent::RequestBonusRefunded { .. } => ctx.notify(),
+                AIRequestUsageModelEvent::AmbientCreditsBannerDismissed => {}
+            }
+            Self::refresh_base_model_menu(&me.base_model_dropdown, ctx);
+            Self::refresh_coding_model_menu(&me.coding_model_dropdown, ctx);
+        });
+
+        let profile_views = Self::create_profile_views(ctx);
+
+        let add_profile_button = ctx.add_typed_action_view(|_| {
+            ActionButton::new("Add Profile", SecondaryTheme)
+                .with_icon(Icon::Plus)
+                .with_size(ButtonSize::Small)
+                .on_click(|ctx| {
+                    ctx.dispatch_typed_action(AgentProfilesPageAction::CreateProfile);
+                })
+        });
+
+        add_profile_button.update(ctx, |button, ctx| {
+            button.set_disabled(!is_any_ai_enabled, ctx);
+        });
+
+        Self {
+            page: Self::build_page(ctx),
+            local_only_icon_tooltip_states: Default::default(),
+            command_execution_allowlist_editor,
+            command_execution_denylist_editor,
+            command_execution_allowlist_mouse_state_handles,
+            command_execution_denylist_mouse_state_handles,
+            base_model_dropdown,
+            coding_model_dropdown,
+            context_window_slider_state,
+            context_window_editor,
+            last_synced_context_window_editor_value,
+            dragged_context_window_value: None,
+            autonomy_dropdown_menu,
+            code_read_allowlist_editor,
+            code_read_autonomy_dropdown_menu,
+            code_read_allowlist_mouse_state_handles,
+            apply_code_diffs_dropdown_menu,
+            read_files_dropdown_menu,
+            execute_commands_dropdown_menu,
+            write_to_pty_autonomy_dropdown_menu,
+            mcp_permissions_dropdown_menu,
+            directory_allowlist_mouse_state_handles,
+            directory_allowlist_editor,
+            command_denylist_mouse_state_handles,
+            command_denylist_tooltip_mouse_state_handles,
+            command_denylist_editor,
+            command_allowlist_mouse_state_handles,
+            command_allowlist_editor,
+            mcp_allowlist_dropdown,
+            mcp_allowlist_mouse_state_handles,
+            mcp_denylist_dropdown,
+            mcp_denylist_mouse_state_handles,
+            profile_views,
+            add_profile_button,
+        }
+    }
+
+    fn build_page(ctx: &mut ViewContext<Self>) -> PageType<Self> {
+        let mut widgets: Vec<Box<dyn SettingsWidget<View = Self>>> = Vec::new();
+        if !FeatureFlag::UsageBasedPricing.is_enabled() {
+            widgets.push(Box::new(UsageWidget::new(ctx)));
+        }
+        widgets.push(Box::new(AgentsWidget::default()));
+
+        // This page is multi-section: it renders its own subheader-sized
+        // section titles inside each widget, so it gets no page-level title.
+        PageType::new_uncategorized(widgets, None)
+    }
+
+    fn handle_context_window_editor_event(
+        &mut self,
+        event: &EditorEvent,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        match event {
+            EditorEvent::Blurred | EditorEvent::Enter => {
+                if !AISettings::as_ref(ctx).is_any_ai_enabled(ctx) {
+                    self.sync_context_window_editor(ctx, true);
+                    return;
+                }
+                if let Some(cw) = Self::configurable_context_window(ctx) {
+                    let buffer_text = self.context_window_editor.as_ref(ctx).buffer_text(ctx);
+                    let cleaned: String = buffer_text
+                        .chars()
+                        .filter(|c| !c.is_whitespace() && *c != ',')
+                        .collect();
+                    if let Ok(parsed) = cleaned.parse::<u32>() {
+                        let clamped = parsed.clamp(cw.min, cw.max);
+                        if Some(clamped) != Self::current_context_window_display_value(ctx) {
+                            AIExecutionProfilesModel::handle(ctx).update(
+                                ctx,
+                                |profiles_model, ctx| {
+                                    let profile_id =
+                                        profiles_model.active_profile(None, ctx).id().clone();
+                                    profiles_model.set_context_window_limit(
+                                        &profile_id,
+                                        Some(clamped),
+                                        ctx,
+                                    );
+                                },
+                            );
+                        }
+                    }
+                }
+                self.sync_context_window_editor(ctx, true);
+                if let EditorEvent::Enter = event {
+                    ctx.emit(AgentProfilesPageEvent::FocusModal);
+                }
+                ctx.notify();
+            }
+            EditorEvent::Escape => ctx.emit(AgentProfilesPageEvent::FocusModal),
+            _ => {}
+        }
+    }
+
+    fn active_profile_data(app: &AppContext) -> AIExecutionProfile {
+        AIExecutionProfilesModel::as_ref(app)
+            .active_profile(None, app)
+            .data()
+            .clone()
+    }
+
+    fn configurable_context_window(app: &AppContext) -> Option<LLMContextWindow> {
+        Self::active_profile_data(app).configurable_context_window(app)
+    }
+
+    fn current_context_window_display_value(app: &AppContext) -> Option<u32> {
+        Self::active_profile_data(app).context_window_display_value(app)
+    }
+
+    fn initial_context_window_value(app: &AppContext) -> u32 {
+        Self::current_context_window_display_value(app).unwrap_or_else(|| {
+            LLMPreferences::as_ref(app)
+                .get_active_base_model(app, None)
+                .context_window
+                .default_max
+        })
+    }
+
+    fn sync_context_window_editor(&mut self, ctx: &mut ViewContext<Self>, force: bool) {
+        self.dragged_context_window_value = None;
+        let Some(value) = Self::current_context_window_display_value(ctx) else {
+            self.last_synced_context_window_editor_value = None;
+            self.context_window_slider_state.reset_offset();
+            ctx.notify();
+            return;
+        };
+
+        let formatted = value.to_string();
+        let should_update = if force {
+            true
+        } else {
+            match self.last_synced_context_window_editor_value {
+                Some(last_value) => {
+                    self.context_window_editor.as_ref(ctx).buffer_text(ctx)
+                        == last_value.to_string()
+                }
+                None => true,
+            }
+        };
+
+        if should_update {
+            self.context_window_editor.update(ctx, |editor, ctx| {
+                if editor.buffer_text(ctx) != formatted {
+                    editor.system_reset_buffer_text(&formatted, ctx);
+                }
+            });
+            self.last_synced_context_window_editor_value = Some(value);
+            self.context_window_slider_state.reset_offset();
+            ctx.notify();
+        }
+    }
+
+    pub fn refresh_base_model_menu(
+        menu: &ViewHandle<Dropdown<AgentProfilesPageAction>>,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        menu.update(ctx, |menu, ctx| {
+            let disabled_by_ai_toggle = !AISettings::as_ref(ctx).is_any_ai_enabled(ctx);
+
+            if disabled_by_ai_toggle {
+                menu.set_disabled(ctx);
+            } else {
+                menu.set_enabled(ctx);
+            }
+
+            let choices = LLMPreferences::as_ref(ctx)
+                .get_base_llm_choices_for_agent_mode(ctx)
+                .collect_vec();
+
+            let items = available_model_menu_items(
+                choices,
+                |llm| {
+                    DropdownAction::select_action_and_close(AgentProfilesPageAction::SetBaseModel(
+                        llm.id.clone(),
+                    ))
+                },
+                None,
+                None,
+                false,
+                false,
+                ctx,
+            );
+            menu.set_rich_items(items, ctx);
+
+            let active = LLMPreferences::as_ref(ctx).get_active_base_model(ctx, None);
+            menu.set_selected_by_action(
+                AgentProfilesPageAction::SetBaseModel(active.id.clone()),
+                ctx,
+            );
+            ctx.notify();
+        });
+        ctx.notify();
+    }
+
+    pub fn refresh_coding_model_menu(
+        menu: &ViewHandle<Dropdown<AgentProfilesPageAction>>,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        menu.update(ctx, |menu, ctx| {
+            let disabled_by_ai_toggle = !AISettings::as_ref(ctx).is_any_ai_enabled(ctx);
+
+            if disabled_by_ai_toggle {
+                menu.set_disabled(ctx);
+            } else {
+                menu.set_enabled(ctx);
+            }
+
+            let choices = LLMPreferences::as_ref(ctx)
+                .get_coding_llm_choices(ctx)
+                .collect_vec();
+
+            let items = available_model_menu_items(
+                choices,
+                |llm| {
+                    DropdownAction::select_action_and_close(
+                        AgentProfilesPageAction::SetCodingModel(llm.id.clone()),
+                    )
+                },
+                None,
+                None,
+                false,
+                false,
+                ctx,
+            );
+            menu.set_rich_items(items, ctx);
+            let active = LLMPreferences::as_ref(ctx).get_active_coding_model(ctx, None);
+
+            menu.set_selected_by_action(
+                AgentProfilesPageAction::SetCodingModel(active.id.clone()),
+                ctx,
+            );
+            ctx.notify();
+        });
+        ctx.notify();
+    }
+
+    fn refresh_autonomy_dropdown_menu(
+        menu: &ViewHandle<Dropdown<AgentProfilesPageAction>>,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        menu.update(ctx, |menu, ctx| {
+            if AISettings::as_ref(ctx).is_any_ai_enabled(ctx) {
+                menu.set_enabled(ctx);
+            } else {
+                menu.set_disabled(ctx);
+            }
+
+            menu.set_items(
+                vec![
+                    DropdownItem::new(
+                        "Read only",
+                        AgentProfilesPageAction::SetAutonomyReadonlyCommandsSetting,
+                    ),
+                    DropdownItem::new(
+                        "Supervised",
+                        AgentProfilesPageAction::SetAutonomySupervisedSetting,
+                    ),
+                ],
+                ctx,
+            );
+            let active = if *AISettings::as_ref(ctx).agent_mode_execute_read_only_commands {
+                0
+            } else {
+                1
+            };
+            menu.set_selected_by_index(active, ctx);
+            ctx.notify();
+        });
+        ctx.notify();
+    }
+
+    fn refresh_all_execution_profile_ui(&self, ctx: &mut ViewContext<Self>) {
+        let permissions = BlocklistAIPermissions::handle(ctx);
+
+        let apply_code_diffs_setting = permissions
+            .as_ref(ctx)
+            .get_apply_code_diffs_setting(ctx, None);
+        Self::refresh_execution_profile_dropdown_menu(
+            &self.apply_code_diffs_dropdown_menu,
+            apply_code_diffs_setting,
+            !AISettings::as_ref(ctx).is_code_diffs_permissions_editable(ctx),
+            ctx,
+        );
+
+        let read_files_setting = permissions.as_ref(ctx).get_read_files_setting(ctx, None);
+        Self::refresh_execution_profile_dropdown_menu(
+            &self.read_files_dropdown_menu,
+            read_files_setting,
+            !AISettings::as_ref(ctx).is_read_files_permissions_editable(ctx),
+            ctx,
+        );
+
+        let execute_commands_setting: ActionPermission = permissions
+            .as_ref(ctx)
+            .get_execute_commands_setting(ctx, None);
+        Self::refresh_execution_profile_dropdown_menu(
+            &self.execute_commands_dropdown_menu,
+            execute_commands_setting,
+            !AISettings::as_ref(ctx).is_execute_commands_permissions_editable(ctx),
+            ctx,
+        );
+
+        let write_to_pty_setting: WriteToPtyPermission =
+            permissions.as_ref(ctx).get_write_to_pty_setting(ctx, None);
+        Self::refresh_write_to_pty_dropdown_menu(
+            &self.write_to_pty_autonomy_dropdown_menu,
+            write_to_pty_setting,
+            !AISettings::as_ref(ctx).is_write_to_pty_permissions_editable(ctx),
+            ctx,
+        );
+
+        let mcp_permissions_setting = permissions
+            .as_ref(ctx)
+            .get_mcp_permissions_setting(ctx, None);
+        Self::refresh_execution_profile_dropdown_menu(
+            &self.mcp_permissions_dropdown_menu,
+            mcp_permissions_setting,
+            !AISettings::as_ref(ctx).is_mcp_permission_editable(ctx),
+            ctx,
+        );
+        Self::refresh_mcp_allowlist_dropdown(&self.mcp_allowlist_dropdown, ctx);
+        Self::refresh_mcp_denylist_dropdown(&self.mcp_denylist_dropdown, ctx);
+
+        let is_any_ai_enabled = AISettings::as_ref(ctx).is_any_ai_enabled(ctx);
+        self.add_profile_button.update(ctx, |button, ctx| {
+            button.set_disabled(!is_any_ai_enabled, ctx);
+        });
+    }
+
+    fn reset_execution_profile_mouse_state_handles(&mut self, ctx: &mut ViewContext<Self>) {
+        let blocklist_permissions = BlocklistAIPermissions::as_ref(ctx);
+
+        self.directory_allowlist_mouse_state_handles = blocklist_permissions
+            .get_read_files_allowlist(ctx, None)
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        self.command_denylist_mouse_state_handles = blocklist_permissions
+            .get_execute_commands_denylist(ctx, None)
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        let org_denylist = BlocklistAIPermissions::get_org_execute_commands_denylist(ctx);
+        self.command_denylist_tooltip_mouse_state_handles =
+            org_denylist.iter().map(|_| Default::default()).collect();
+
+        self.command_allowlist_mouse_state_handles = blocklist_permissions
+            .get_execute_commands_allowlist(ctx, None)
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        self.mcp_allowlist_mouse_state_handles = blocklist_permissions
+            .get_mcp_allowlist(ctx, None)
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+
+        self.mcp_denylist_mouse_state_handles = blocklist_permissions
+            .get_mcp_denylist(ctx, None)
+            .iter()
+            .map(|_| Default::default())
+            .collect();
+    }
+
+    fn refresh_execution_profile_dropdown_menu(
+        menu: &ViewHandle<Dropdown<AgentProfilesPageAction>>,
+        current_permission: ActionPermission,
+        disabled: bool,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        menu.update(ctx, |menu, ctx| {
+            if !disabled {
+                menu.set_enabled(ctx);
+            } else {
+                menu.set_disabled(ctx);
+            }
+
+            let active = match current_permission {
+                ActionPermission::AgentDecides | ActionPermission::Unknown => 0,
+                ActionPermission::AlwaysAllow => 1,
+                ActionPermission::AlwaysAsk => 2,
+            };
+
+            menu.set_selected_by_index(active, ctx);
+            ctx.notify();
+        });
+        ctx.notify();
+    }
+
+    fn refresh_write_to_pty_dropdown_menu(
+        menu: &ViewHandle<Dropdown<AgentProfilesPageAction>>,
+        current_permission: WriteToPtyPermission,
+        disabled: bool,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        menu.update(ctx, |menu, ctx| {
+            if !disabled {
+                menu.set_enabled(ctx);
+            } else {
+                menu.set_disabled(ctx);
+            }
+
+            let active = match current_permission {
+                WriteToPtyPermission::AlwaysAllow => 0,
+                WriteToPtyPermission::AlwaysAsk | WriteToPtyPermission::Unknown => 1,
+                WriteToPtyPermission::AskOnFirstWrite => 2,
+            };
+
+            menu.set_selected_by_index(active, ctx);
+            ctx.notify();
+        });
+        ctx.notify();
+    }
+
+    /// Currently, the coding permissions only support "read" access.
+    fn refresh_code_read_autonomy_dropdown_menu(
+        menu: &ViewHandle<Dropdown<AgentProfilesPageAction>>,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        menu.update(ctx, |menu, ctx| {
+            if AISettings::as_ref(ctx).is_any_ai_enabled(ctx) {
+                menu.set_enabled(ctx);
+            } else {
+                menu.set_disabled(ctx);
+            }
+
+            menu.set_items(
+                AgentModeCodingPermissionsType::iter()
+                    .map(|t| {
+                        let display = match t {
+                            AgentModeCodingPermissionsType::AlwaysAskBeforeReading => "Always ask",
+                            AgentModeCodingPermissionsType::AlwaysAllowReading => "Always allow",
+                            AgentModeCodingPermissionsType::AllowReadingSpecificFiles => {
+                                "Allow in specific directories"
+                            }
+                        };
+                        DropdownItem::new(display, AgentProfilesPageAction::SetCodingPermission(t))
+                    })
+                    .collect(),
+                ctx,
+            );
+            let ai_settings = AISettings::as_ref(ctx);
+
+            let active = if *ai_settings.agent_mode_execute_read_only_commands {
+                menu.set_disabled(ctx);
+                AgentModeCodingPermissionsType::AlwaysAllowReading
+            } else {
+                *ai_settings.agent_mode_coding_permissions
+            };
+            menu.set_selected_by_action(AgentProfilesPageAction::SetCodingPermission(active), ctx);
+            ctx.notify();
+        });
+        ctx.notify();
+    }
+
+    fn get_non_allowlisted_or_denylisted_mcp_servers(
+        ctx: &mut ViewContext<Self>,
+    ) -> Vec<(uuid::Uuid, String)> {
+        let all_mcp_servers = TemplatableMCPServerManager::get_all_cloud_synced_mcp_servers(ctx);
+        let already_allowlisted_mcp_servers =
+            BlocklistAIPermissions::as_ref(ctx).get_mcp_allowlist(ctx, None);
+        let already_denylisted_mcp_servers =
+            BlocklistAIPermissions::as_ref(ctx).get_mcp_denylist(ctx, None);
+
+        all_mcp_servers
+            .into_iter()
+            .filter(|(uuid, _)| {
+                let is_allowlisted = already_allowlisted_mcp_servers.contains(uuid);
+                let is_denylisted = already_denylisted_mcp_servers.contains(uuid);
+                !is_allowlisted && !is_denylisted
+            })
+            .collect()
+    }
+
+    fn refresh_menu_dropdown<F>(
+        menu: &ViewHandle<FilterableDropdown<AgentProfilesPageAction>>,
+        action_fn: F,
+        ctx: &mut ViewContext<Self>,
+    ) where
+        F: Fn(uuid::Uuid) -> AgentProfilesPageAction,
+    {
+        let mcps_in_dropdown = Self::get_non_allowlisted_or_denylisted_mcp_servers(ctx);
+        menu.update(ctx, |menu, ctx| {
+            if AISettings::as_ref(ctx).is_any_ai_enabled(ctx) {
+                menu.set_enabled(ctx);
+            } else {
+                menu.set_disabled(ctx);
+            }
+
+            let items: Vec<DropdownItem<AgentProfilesPageAction>> = mcps_in_dropdown
+                .iter()
+                .map(|(uuid, server_name)| DropdownItem::new(server_name, action_fn(*uuid)))
+                .collect();
+
+            menu.set_items(items, ctx);
+            ctx.notify()
+        });
+        ctx.notify();
+    }
+
+    fn refresh_mcp_allowlist_dropdown(
+        menu: &ViewHandle<FilterableDropdown<AgentProfilesPageAction>>,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        Self::refresh_menu_dropdown(menu, AgentProfilesPageAction::AddToMCPAllowlist, ctx);
+    }
+
+    fn create_profile_views(ctx: &mut ViewContext<Self>) -> Vec<ViewHandle<ExecutionProfileView>> {
+        let profiles_model = AIExecutionProfilesModel::as_ref(ctx);
+        let profile_ids = profiles_model.get_all_profile_ids();
+
+        profile_ids
+            .iter()
+            .map(|profile_id| {
+                let profile_id = profile_id.clone();
+                let profile_view = ctx.add_typed_action_view(|ctx| {
+                    ExecutionProfileView::new(profile_id.clone(), ctx)
+                });
+                let profile_id_for_event = profile_id.clone();
+
+                ctx.subscribe_to_view(&profile_view, move |_me, _, event, ctx| match event {
+                    ExecutionProfileViewEvent::EditProfile => {
+                        ctx.emit(AgentProfilesPageEvent::OpenExecutionProfileEditor(
+                            profile_id_for_event.clone(),
+                        ));
+                    }
+                });
+
+                profile_view
+            })
+            .collect()
+    }
+
+    fn refresh_profile_views(&mut self, ctx: &mut ViewContext<Self>) {
+        let new_profile_views = Self::create_profile_views(ctx);
+        self.profile_views = new_profile_views;
+    }
+
+    fn refresh_mcp_denylist_dropdown(
+        menu: &ViewHandle<FilterableDropdown<AgentProfilesPageAction>>,
+        ctx: &mut ViewContext<Self>,
+    ) {
+        Self::refresh_menu_dropdown(menu, AgentProfilesPageAction::AddToMCPDenylist, ctx);
+    }
+}
+
+impl View for AgentProfilesPageView {
+    fn ui_name() -> &'static str {
+        "AgentProfilesPage"
+    }
+
+    fn render(&self, app: &AppContext) -> Box<dyn Element> {
+        self.page.render(self, app)
+    }
+}
+
+pub enum AgentProfilesPageEvent {
+    FocusModal,
+    OpenMCPServerCollection,
+    OpenExecutionProfileEditor(ExecutionProfileId),
+}
+
+impl Entity for AgentProfilesPageView {
+    type Event = AgentProfilesPageEvent;
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum AgentProfilesPageAction {
+    HyperlinkClick(HyperlinkUrl),
+    ToggleCodebaseContext,
+    AttemptLoginGatedUpgrade,
+    RemoveFromCommandExecutionAllowlist(AgentModeCommandExecutionPredicate),
+    RemoveFromCommandExecutionDenylist(AgentModeCommandExecutionPredicate),
+    OpenMCPServerCollection,
+    OpenExecutionProfileEditor(ExecutionProfileId),
+    SetBaseModel(LLMId),
+    SetCodingModel(LLMId),
+    /// Called while the user is actively dragging the context window slider.
+    ContextWindowSliderDragged(u32),
+    /// Called when the user commits a new context window value (slider drop or
+    /// input box commit).
+    SetContextWindowSize(u32),
+    SetAutonomyReadonlyCommandsSetting,
+    SetAutonomySupervisedSetting,
+    SetCodingPermission(AgentModeCodingPermissionsType),
+    SetApplyCodeDiffs(ActionPermission),
+    SetReadFiles(ActionPermission),
+    SetExecuteCommands(ActionPermission),
+    SetWriteToPty(WriteToPtyPermission),
+    SetMCPPermissions(ActionPermission),
+    RemoveDirectoryFromCodeReadAllowlist(PathBuf),
+    RemoveFromProfileDirectoryAllowlist(PathBuf),
+    RemoveFromProfileCommandDenylist(AgentModeCommandExecutionPredicate),
+    RemoveFromProfileCommandAllowlist(AgentModeCommandExecutionPredicate),
+    ToggleShowBaseModelPickerInPrompt,
+    AddToMCPAllowlist(uuid::Uuid),
+    RemoveFromMCPAllowlist(uuid::Uuid),
+    AddToMCPDenylist(uuid::Uuid),
+    RemoveFromMCPDenylist(uuid::Uuid),
+    CreateProfile,
+}
+
+impl From<&AgentProfilesPageAction> for LoginGatedFeature {
+    fn from(val: &AgentProfilesPageAction) -> LoginGatedFeature {
+        use AgentProfilesPageAction::*;
+        match val {
+            AttemptLoginGatedUpgrade => "Upgrade AI Usage",
+            _ => "Unknown reason",
+        }
+    }
+}
+
+impl TypedActionView for AgentProfilesPageView {
+    type Action = AgentProfilesPageAction;
+
+    fn handle_action(&mut self, action: &Self::Action, ctx: &mut ViewContext<Self>) {
+        match action {
+            AgentProfilesPageAction::ToggleCodebaseContext => {
+                match CodeSettings::handle(ctx).update(ctx, |settings, ctx| {
+                    settings.codebase_context_enabled.toggle_and_save_value(ctx)
+                }) {
+                    Ok(new_value) => {
+                        send_telemetry_from_ctx!(
+                            TelemetryEvent::ToggleCodebaseContext {
+                                is_codebase_context_enabled: new_value
+                            },
+                            ctx
+                        );
+                    }
+                    Err(e) => {
+                        log::warn!("Failed to set value for Codebase Context: {e:?}");
+                    }
+                }
+                ctx.notify();
+            }
+            AgentProfilesPageAction::HyperlinkClick(hyperlink) => {
+                open_hyperlink(hyperlink, ctx);
+            }
+            AgentProfilesPageAction::AttemptLoginGatedUpgrade => {
+                AuthManager::handle(ctx).update(ctx, |auth_manager, ctx| {
+                    auth_manager.attempt_login_gated_feature(
+                        action.into(),
+                        AuthViewVariant::RequireLoginCloseable,
+                        ctx,
+                    )
+                });
+            }
+            AgentProfilesPageAction::RemoveFromCommandExecutionAllowlist(cmd) => {
+                BlocklistAIPermissions::handle(ctx).update(ctx, |model, ctx| {
+                    report_if_error!(model.remove_command_from_autoexecution_allowlist(cmd, ctx));
+                })
+            }
+            AgentProfilesPageAction::RemoveFromCommandExecutionDenylist(cmd) => {
+                BlocklistAIPermissions::handle(ctx).update(ctx, |model, ctx| {
+                    report_if_error!(model.remove_command_from_denylist(cmd, ctx));
+                })
+            }
+            AgentProfilesPageAction::OpenMCPServerCollection => {
+                ctx.emit(AgentProfilesPageEvent::OpenMCPServerCollection)
+            }
+            AgentProfilesPageAction::OpenExecutionProfileEditor(profile_id) => ctx.emit(
+                AgentProfilesPageEvent::OpenExecutionProfileEditor(profile_id.clone()),
+            ),
+            AgentProfilesPageAction::SetBaseModel(id) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |profiles_model, ctx| {
+                    let profile_id = profiles_model.active_profile(None, ctx).id().clone();
+                    profiles_model.set_base_model(&profile_id, Some(id.clone()), ctx);
+                    profiles_model.set_context_window_limit(&profile_id, None, ctx);
+                });
+                self.sync_context_window_editor(ctx, true);
+                ctx.notify();
+            }
+            AgentProfilesPageAction::SetCodingModel(id) => {
+                LLMPreferences::handle(ctx).update(ctx, |prefs, ctx| {
+                    prefs.update_preferred_coding_llm(id, None, ctx);
+                });
+            }
+            AgentProfilesPageAction::ContextWindowSliderDragged(value) => {
+                if !AISettings::as_ref(ctx).is_any_ai_enabled(ctx) {
+                    self.sync_context_window_editor(ctx, true);
+                    return;
+                }
+                if Self::configurable_context_window(ctx).is_some() {
+                    self.dragged_context_window_value = Some(*value);
+                    let formatted = value.to_string();
+                    self.context_window_editor.update(ctx, |editor, ctx| {
+                        editor.system_reset_buffer_text(&formatted, ctx);
+                    });
+                    ctx.notify();
+                }
+            }
+            AgentProfilesPageAction::SetContextWindowSize(value) => {
+                self.dragged_context_window_value = None;
+                if !AISettings::as_ref(ctx).is_any_ai_enabled(ctx) {
+                    self.sync_context_window_editor(ctx, true);
+                    return;
+                }
+                let Some(cw) = Self::configurable_context_window(ctx) else {
+                    return;
+                };
+                let clamped = (*value).clamp(cw.min, cw.max);
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |profiles_model, ctx| {
+                    let profile_id = profiles_model.active_profile(None, ctx).id().clone();
+                    profiles_model.set_context_window_limit(&profile_id, Some(clamped), ctx);
+                });
+                self.sync_context_window_editor(ctx, true);
+                ctx.notify();
+            }
+            AgentProfilesPageAction::SetAutonomyReadonlyCommandsSetting
+            | AgentProfilesPageAction::SetAutonomySupervisedSetting => {
+                let readonly_cmd_execution_enabled = matches!(
+                    action,
+                    AgentProfilesPageAction::SetAutonomyReadonlyCommandsSetting
+                );
+                BlocklistAIPermissions::handle(ctx).update(ctx, |model, ctx| {
+                    match model.set_should_autoexecute_readonly_commands(
+                        readonly_cmd_execution_enabled,
+                        ctx,
+                    ) {
+                        Ok(_) => {
+                            send_telemetry_from_ctx!(
+                                TelemetryEvent::ToggledAgentModeAutoexecuteReadonlyCommandsSetting {
+                                    src: AutonomySettingToggleSource::SettingsPage,
+                                    enabled: readonly_cmd_execution_enabled,
+                                },
+                                ctx);
+                        }
+                        Err(e) => report_error!(e),
+                    }
+                });
+            }
+            AgentProfilesPageAction::SetCodingPermission(p) => {
+                BlocklistAIPermissions::handle(ctx).update(ctx, |model, ctx| {
+                    match model.set_coding_permissions(*p, ctx) {
+                        Ok(_) => {
+                            send_telemetry_from_ctx!(
+                                TelemetryEvent::ChangedAgentModeCodingPermissions {
+                                    src: AutonomySettingToggleSource::SettingsPage,
+                                    new: *p,
+                                },
+                                ctx
+                            );
+                        }
+                        Err(e) => report_error!(e),
+                    }
+                });
+            }
+            AgentProfilesPageAction::SetApplyCodeDiffs(permission) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    model.set_apply_code_diffs(profile.id(), permission, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::SetReadFiles(permission) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    model.set_read_files(profile.id(), permission, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::SetExecuteCommands(permission) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    model.set_execute_commands(profile.id(), permission, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::SetWriteToPty(permission) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    model.set_write_to_pty(profile.id(), permission, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::SetMCPPermissions(permission) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    model.set_mcp_permissions(profile.id(), permission, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::RemoveDirectoryFromCodeReadAllowlist(dir) => {
+                BlocklistAIPermissions::handle(ctx).update(ctx, |model, ctx| {
+                    report_if_error!(
+                        model.remove_filepath_from_code_read_allowlist(dir.to_owned(), ctx)
+                    );
+                });
+            }
+            AgentProfilesPageAction::RemoveFromProfileDirectoryAllowlist(path_buf) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+                    model.remove_from_directory_allowlist(
+                        profile_id,
+                        &PathBuf::from(path_buf),
+                        ctx,
+                    );
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::RemoveFromProfileCommandDenylist(cmd) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+
+                    model.remove_from_command_denylist(profile_id, cmd, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::RemoveFromProfileCommandAllowlist(command) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+
+                    model.remove_from_command_allowlist(profile_id, command, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::ToggleShowBaseModelPickerInPrompt => {
+                SessionSettings::handle(ctx).update(ctx, |settings, ctx| {
+                    if let Err(e) = settings
+                        .show_model_selectors_in_prompt
+                        .toggle_and_save_value(ctx)
+                    {
+                        log::warn!(
+                            "Failed to set value for Show Base Model Picker in Prompt: {e:?}"
+                        );
+                    }
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::AddToMCPAllowlist(id) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+                    model.add_to_mcp_allowlist(profile_id, id, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::RemoveFromMCPAllowlist(id) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+                    model.remove_from_mcp_allowlist(profile_id, id, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::AddToMCPDenylist(id) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+                    model.add_to_mcp_denylist(profile_id, id, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::RemoveFromMCPDenylist(id) => {
+                AIExecutionProfilesModel::handle(ctx).update(ctx, |model, ctx| {
+                    let profile = model.default_profile(ctx);
+                    let profile_id = profile.id();
+                    model.remove_from_mcp_denylist(profile_id, id, ctx);
+                });
+                ctx.notify();
+            }
+            AgentProfilesPageAction::CreateProfile => {
+                let new_profile_id = AIExecutionProfilesModel::handle(ctx)
+                    .update(ctx, |model, ctx| model.create_profile(ctx));
+
+                if let Some(profile_id) = new_profile_id {
+                    ctx.emit(AgentProfilesPageEvent::OpenExecutionProfileEditor(
+                        profile_id,
+                    ));
+                }
+                ctx.notify();
+            }
+        }
+    }
+}
+
+impl SettingsPageMeta for AgentProfilesPageView {
+    fn section() -> SettingsSection {
+        SettingsSection::AgentProfiles
+    }
+
+    fn should_render(&self, _ctx: &AppContext) -> bool {
+        FeatureFlag::AgentMode.is_enabled()
+    }
+
+    fn on_page_selected(&mut self, _: bool, ctx: &mut ViewContext<Self>) {
+        // The usage widget on this page is the only consumer of the request
+        // usage model, so the refresh lives here.
+        AIRequestUsageModel::handle(ctx).update(ctx, |ai_request_usage_model, ctx| {
+            ai_request_usage_model.refresh_request_usage_async(ctx)
+        });
+    }
+
+    fn update_filter(&mut self, query: &str, ctx: &mut ViewContext<Self>) -> MatchData {
+        self.page.update_filter(query, ctx)
+    }
+
+    fn scroll_to_widget(&mut self, widget_id: &'static str) {
+        self.page.scroll_to_widget(widget_id)
+    }
+
+    fn clear_highlighted_widget(&mut self) {
+        self.page.clear_highlighted_widget();
+    }
+}
+
+impl From<ViewHandle<AgentProfilesPageView>> for SettingsPageViewHandle {
+    fn from(view_handle: ViewHandle<AgentProfilesPageView>) -> Self {
+        SettingsPageViewHandle::AgentProfiles(view_handle)
+    }
+}
+
+pub fn init_actions_from_parent_view<T: Action + Clone>(
+    app: &mut AppContext,
+    context: &ContextPredicate,
+    builder: fn(SettingsAction) -> T,
+) {
+    ToggleSettingActionPair::add_toggle_setting_action_pairs_as_bindings(
+        vec![
+            ToggleSettingActionPair::new(
+                "model picker in prompt",
+                builder(SettingsAction::AgentProfiles(
+                    AgentProfilesPageAction::ToggleShowBaseModelPickerInPrompt,
+                )),
+                &(context.clone() & id!(flags::IS_ANY_AI_ENABLED)),
+                flags::SHOW_BASE_MODEL_PICKER_IN_PROMPT_FLAG,
+            )
+            .with_group(bindings::BindingGroup::WarpAi),
+        ],
+        app,
+    );
+    if !FeatureFlag::FullSourceCodeEmbedding.is_enabled() {
+        ToggleSettingActionPair::add_toggle_setting_action_pairs_as_bindings(
+            vec![ToggleSettingActionPair::new(
+                "codebase index",
+                builder(SettingsAction::AgentProfiles(
+                    AgentProfilesPageAction::ToggleCodebaseContext,
+                )),
+                &(context.clone() & id!(flags::IS_ANY_AI_ENABLED)),
+                flags::IS_CODEBASE_INDEXING_ENABLED,
+            )],
+            app,
+        );
+    }
+}
+
+fn render_ai_list(
+    header: &str,
+    description: &str,
+    input_list: Box<dyn Element>,
+    view: &AgentProfilesPageView,
+    ai_settings: &AISettings,
+    app: &AppContext,
+) -> Box<dyn Element> {
+    let setting_header = render_ai_setting_label::<AgentModeCommandExecutionDenylist>(
+        header.to_string(),
+        ai_settings.is_any_ai_enabled(app),
+        &view.local_only_icon_tooltip_states,
+        app,
+    );
+
+    let description = render_ai_setting_description(
+        description.to_string(),
+        ai_settings.is_any_ai_enabled(app),
+        app,
+    );
+
+    Flex::column()
+        .with_child(setting_header)
+        .with_child(Container::new(description).with_margin_bottom(-8.).finish())
+        .with_child(input_list)
+        .finish()
+}
+
+struct UsageWidget {
+    view_handle: WeakViewHandle<AgentProfilesPageView>,
+    requests_highlight_index: HighlightedHyperlink,
+}
+
+impl UsageWidget {
+    fn new(ctx: &ViewContext<AgentProfilesPageView>) -> Self {
+        Self {
+            view_handle: ctx.handle(),
+            requests_highlight_index: Default::default(),
+        }
+    }
+    fn render_request_usage_count(
+        &self,
+        used: usize,
+        limit: usize,
+        is_unlimited: bool,
+        workspace_is_delinquent_due_to_payment_issue: bool,
+        appearance: &Appearance,
+    ) -> Box<dyn warpui::Element> {
+        let mut row = Flex::row();
+        if used >= limit || workspace_is_delinquent_due_to_payment_issue {
+            row.add_child(
+                ConstrainedBox::new(
+                    Icon::AlertTriangle
+                        .to_warpui_icon(appearance.theme().ui_error_color().into())
+                        .finish(),
+                )
+                .with_height(16.)
+                .with_width(16.)
+                .finish(),
+            )
+        }
+
+        let request_count_label = if workspace_is_delinquent_due_to_payment_issue {
+            "Restricted due to billing issue".to_string()
+        } else if is_unlimited {
+            "Unlimited".to_string()
+        } else {
+            format!("{used}/{limit}")
+        };
+
+        row.add_child(
+            appearance
+                .ui_builder()
+                .paragraph(request_count_label)
+                .with_style(UiComponentStyles {
+                    font_color: {
+                        if used >= limit {
+                            Some(appearance.theme().ui_error_color())
+                        } else {
+                            Some(blended_colors::text_sub(
+                                appearance.theme(),
+                                appearance.theme().surface_1(),
+                            ))
+                        }
+                    },
+                    font_size: Some(16.),
+                    margin: Some(Coords {
+                        top: 0.,
+                        bottom: 0.,
+                        left: 8.,
+                        right: 0.,
+                    }),
+                    ..Default::default()
+                })
+                .build()
+                .finish(),
+        );
+
+        row.finish()
+    }
+
+    /// Renders a row of what is being limited, along with the current used/limit.
+    #[allow(clippy::too_many_arguments)]
+    fn render_ai_usage_limit_row(
+        &self,
+        header: impl Into<Cow<'static, str>>,
+        description: impl Into<Cow<'static, str>>,
+        used: usize,
+        limit: usize,
+        is_unlimited: bool,
+        workspace_is_delinquent_due_to_payment_issue: bool,
+        appearance: &Appearance,
+    ) -> Box<dyn warpui::Element> {
+        let request_usage_details = Flex::column()
+            .with_cross_axis_alignment(CrossAxisAlignment::End)
+            .with_child(self.render_request_usage_count(
+                used,
+                limit,
+                is_unlimited,
+                workspace_is_delinquent_due_to_payment_issue,
+                appearance,
+            ));
+
+        let request_usage_description = FormattedTextElement::from_str(
+            description,
+            appearance.ui_font_family(),
+            appearance.ui_font_size(),
+        )
+        .with_color(blended_colors::text_sub(
+            appearance.theme(),
+            appearance.theme().surface_1(),
+        ));
+
+        Flex::row()
+            .with_child(
+                Shrinkable::new(
+                    2.,
+                    Container::new(
+                        Flex::column()
+                            .with_child(
+                                appearance
+                                    .ui_builder()
+                                    .paragraph(header)
+                                    .with_style(UiComponentStyles {
+                                        font_color: Some(blended_colors::text_main(
+                                            appearance.theme(),
+                                            appearance.theme().surface_1(),
+                                        )),
+                                        margin: Some(Coords {
+                                            top: 0.,
+                                            bottom: 4.,
+                                            left: 0.,
+                                            right: 0.,
+                                        }),
+                                        ..Default::default()
+                                    })
+                                    .build()
+                                    .finish(),
+                            )
+                            .with_child(request_usage_description.finish())
+                            .finish(),
+                    )
+                    .with_margin_bottom(16.)
+                    .finish(),
+                )
+                .finish(),
+            )
+            .with_child(
+                Shrinkable::new(
+                    1.,
+                    Container::new(request_usage_details.finish())
+                        .with_margin_bottom(16.)
+                        .finish(),
+                )
+                .finish(),
+            )
+            .with_cross_axis_alignment(CrossAxisAlignment::Center)
+            .with_main_axis_alignment(MainAxisAlignment::SpaceBetween)
+            .with_main_axis_size(MainAxisSize::Max)
+            .finish()
+    }
+}
+
+impl SettingsWidget for UsageWidget {
+    type View = AgentProfilesPageView;
+
+    fn search_terms(&self) -> &str {
+        "a.i. ai usage limit plan"
+    }
+
+    fn render(
+        &self,
+        _view: &Self::View,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let ai_request_usage_model = AIRequestUsageModel::as_ref(app);
+        let next_refresh_time = ai_request_usage_model.next_refresh_time();
+        let formatted_next_refresh_time = next_refresh_time.format("%b %d").to_string();
+        let workspace_is_delinquent_due_to_payment_issue = UserWorkspaces::as_ref(app)
+            .team_for_view_handle(&self.view_handle, app)
+            .map(|team| team.billing_metadata.is_delinquent_due_to_payment_issue())
+            .unwrap_or_default();
+
+        let usage_header = Container::new(
+            Flex::row()
+                .with_main_axis_size(MainAxisSize::Max)
+                .with_main_axis_alignment(MainAxisAlignment::SpaceBetween)
+                .with_cross_axis_alignment(CrossAxisAlignment::Center)
+                .with_child(
+                    build_sub_header(
+                        appearance,
+                        "Usage",
+                        Some(styles::header_font_color(true, app)),
+                    )
+                    .finish(),
+                )
+                .with_child(
+                    appearance
+                        .ui_builder()
+                        .paragraph(format!("Resets {formatted_next_refresh_time}"))
+                        .with_style(UiComponentStyles {
+                            font_color: Some(blended_colors::text_sub(
+                                appearance.theme(),
+                                appearance.theme().surface_1(),
+                            )),
+                            ..Default::default()
+                        })
+                        .build()
+                        .finish(),
+                )
+                .finish(),
+        )
+        .with_padding_bottom(HEADER_PADDING)
+        .finish();
+
+        let request_limit_description = format!(
+            "This is the {} limit of AI credits for your account.",
+            ai_request_usage_model.refresh_duration_to_string()
+        );
+
+        let request_usage_row = self.render_ai_usage_limit_row(
+            "Credits",
+            request_limit_description,
+            ai_request_usage_model.requests_used(),
+            ai_request_usage_model.request_limit(),
+            ai_request_usage_model.is_unlimited(),
+            workspace_is_delinquent_due_to_payment_issue,
+            appearance,
+        );
+
+        let auth_state = AuthStateProvider::as_ref(app).get();
+        let upgrade_cta_text_fragments = if let Some(team) =
+            UserWorkspaces::as_ref(app).team_for_view_handle(&self.view_handle, app)
+        {
+            let current_user_email = auth_state.user_email().unwrap_or_default();
+            let has_admin_permissions = team.has_admin_permissions(&current_user_email);
+            if team.billing_metadata.can_upgrade_to_higher_tier_plan() {
+                let upgrade_url = UserWorkspaces::upgrade_link_for_team(team.uid);
+                if has_admin_permissions {
+                    vec![
+                        FormattedTextFragment::hyperlink("Upgrade", upgrade_url),
+                        FormattedTextFragment::plain_text(" to get more AI usage."),
+                    ]
+                } else {
+                    // The /upgrade page says to contact their administrator.
+                    vec![
+                        FormattedTextFragment::hyperlink("Compare plans", upgrade_url),
+                        FormattedTextFragment::plain_text(" for more AI usage."),
+                    ]
+                }
+            } else {
+                vec![
+                    FormattedTextFragment::hyperlink("Contact support", "mailto:support@warp.dev"),
+                    FormattedTextFragment::plain_text(" for more AI usage."),
+                ]
+            }
+        } else {
+            let user_id = auth_state.user_id().unwrap_or_default();
+            let upgrade_url = UserWorkspaces::upgrade_link(user_id);
+            vec![
+                FormattedTextFragment::hyperlink("Upgrade", upgrade_url),
+                FormattedTextFragment::plain_text(" to get more AI usage."),
+            ]
+        };
+
+        let mut upgrade_cta = FormattedTextElement::new(
+            FormattedText::new([FormattedTextLine::Line(upgrade_cta_text_fragments)]),
+            appearance.ui_font_size(),
+            appearance.ui_font_family(),
+            appearance.ui_font_family(),
+            styles::description_font_color(true, app).into(),
+            self.requests_highlight_index.clone(),
+        )
+        .with_hyperlink_font_color(appearance.theme().accent().into_solid());
+
+        if AuthStateProvider::as_ref(app)
+            .get()
+            .is_anonymous_or_logged_out()
+        {
+            upgrade_cta = upgrade_cta.register_default_click_handlers(|_, ctx, _| {
+                ctx.dispatch_typed_action(AgentProfilesPageAction::AttemptLoginGatedUpgrade);
+            });
+        } else {
+            upgrade_cta = upgrade_cta.register_default_click_handlers(|url, ctx, _| {
+                ctx.dispatch_typed_action(AgentProfilesPageAction::HyperlinkClick(url));
+            })
+        }
+
+        Flex::column()
+            .with_children([
+                render_separator(appearance),
+                usage_header,
+                request_usage_row,
+                Container::new(upgrade_cta.finish())
+                    .with_margin_bottom(16.)
+                    .finish(),
+            ])
+            .finish()
+    }
+}
+
+#[derive(Default)]
+struct AgentsWidget {
+    codebase_context_toggle: SwitchStateHandle,
+    codebase_context_link_index: HighlightedHyperlink,
+    show_in_prompt_checkbox: MouseStateHandle,
+}
+
+impl SettingsWidget for AgentsWidget {
+    type View = AgentProfilesPageView;
+
+    fn search_terms(&self) -> &str {
+        if should_show_mcp_servers() {
+            "ai a.i. agent autonomy profiles allowlist denylist autoexecute permissions models llms planning mcp server"
+        } else {
+            "ai a.i. agent autonomy profiles allowlist denylist autoexecute permissions models llms planning"
+        }
+    }
+
+    fn render(
+        &self,
+        view: &Self::View,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let ai_settings = AISettings::as_ref(app);
+        let is_any_ai_enabled = ai_settings.is_any_ai_enabled(app);
+
+        let mut column = Flex::column();
+
+        if FeatureFlag::ProfilesDesignRevamp.is_enabled() {
+            column.add_child(
+                Container::new(self.render_profiles_section(view, ai_settings, appearance, app))
+                    .with_margin_bottom(8.)
+                    .finish(),
+            );
+        } else {
+            // Legacy layout: show Agents header + Models + Permissions
+            let mut agents_header = Flex::column();
+            agents_header.add_child(
+                build_sub_header(
+                    appearance,
+                    "Agents",
+                    Some(styles::header_font_color(is_any_ai_enabled, app)),
+                )
+                .with_padding_bottom(HEADER_PADDING)
+                .finish(),
+            );
+            agents_header.add_child(render_ai_setting_description(
+                "Set the boundaries for how your Agent operates. Choose what it can access, how much autonomy it has, and when it must ask for your approval. You can also fine-tune behavior around natural language input, codebase awareness, and more.",
+                ai_settings.is_any_ai_enabled(app),
+                app,
+            ));
+            let agents_header = agents_header.finish();
+            column.add_children([
+                render_separator(appearance),
+                Container::new(agents_header)
+                    .with_margin_bottom(8.)
+                    .finish(),
+            ]);
+            column.add_children([
+                Container::new(self.render_models_section(view, ai_settings, appearance, app))
+                    .with_margin_bottom(8.)
+                    .finish(),
+                Container::new(self.render_permissions_section(view, ai_settings, appearance, app))
+                    .with_margin_bottom(8.)
+                    .finish(),
+            ]);
+        };
+
+        column.finish()
+    }
+}
+
+impl AgentsWidget {
+    fn render_profiles_section(
+        &self,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let is_any_ai_enabled = ai_settings.is_any_ai_enabled(app);
+
+        let header_and_description = Flex::column()
+            .with_child(
+                build_sub_header(
+                    appearance,
+                    "Profiles",
+                    Some(styles::header_font_color(is_any_ai_enabled, app)),
+                )
+                .finish(),
+            )
+            .with_child(
+                Container::new(
+                    render_ai_setting_description(
+                        "Profiles let you define how your Agent operates — from the actions it can take and when it needs approval, to the models it uses for tasks like coding and planning. You can also scope them to individual projects.",
+                        is_any_ai_enabled,
+                        app,
+                    )
+                )
+                .with_margin_top(12.)
+                .finish()
+            )
+            .finish();
+
+        let mut profiles_row = Flex::row()
+            .with_main_axis_size(MainAxisSize::Max)
+            .with_cross_axis_alignment(CrossAxisAlignment::Start)
+            .with_child(Shrinkable::new(1., header_and_description).finish());
+
+        if FeatureFlag::MultiProfile.is_enabled() {
+            profiles_row.add_child(
+                Container::new(view.add_profile_button.as_ref(app).render(app))
+                    .with_margin_left(16.)
+                    .finish(),
+            );
+        }
+
+        let profiles_header = Container::new(profiles_row.finish())
+            .with_margin_bottom(12.0)
+            .finish();
+
+        let mut profile_elements = vec![profiles_header];
+
+        for profile_view in &view.profile_views {
+            profile_elements.push(
+                Container::new(ChildView::new(profile_view).finish())
+                    .with_margin_bottom(8.)
+                    .finish(),
+            );
+        }
+
+        Flex::column().with_children(profile_elements).finish()
+    }
+
+    fn render_models_section(
+        &self,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let is_any_ai_enabled = ai_settings.is_any_ai_enabled(app);
+        let model_subheader = Container::new(render_custom_size_header(
+            appearance,
+            "Models",
+            14.0,
+            Some(styles::header_font_color(is_any_ai_enabled, app)),
+        ))
+        .with_margin_bottom(8.0)
+        .finish();
+
+        let base_model_setting =
+            Container::new(self.render_base_model_setting(view, ai_settings, appearance, app))
+                .with_margin_bottom(8.0)
+                .finish();
+
+        let mut children = vec![model_subheader, base_model_setting];
+        if let Some(context_window_setting) =
+            self.render_context_window_setting(view, ai_settings, appearance, app)
+        {
+            children.push(
+                Container::new(context_window_setting)
+                    .with_margin_bottom(8.0)
+                    .finish(),
+            );
+        }
+
+        Flex::column().with_children(children).finish()
+    }
+
+    /// Renders the context window slider + numeric input row shown below the
+    /// base model dropdown. Returns `None` if the active base model does not
+    /// advertise a configurable context window or global AI is disabled.
+    fn render_context_window_setting(
+        &self,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Option<Box<dyn Element>> {
+        if !ai_settings.is_any_ai_enabled(app) {
+            return None;
+        }
+        let cw = AgentProfilesPageView::configurable_context_window(app)?;
+        let min = cw.min;
+        let max = cw.max;
+
+        let label = Container::new(render_body_item_label::<AgentProfilesPageAction>(
+            "Context window (tokens)".to_string(),
+            None,
+            None,
+            LocalOnlyIconState::Hidden,
+            ToggleState::Enabled,
+            appearance,
+        ))
+        .with_margin_bottom(4.0)
+        .finish();
+
+        let min_label = appearance
+            .ui_builder()
+            .span(format!("{min}"))
+            .with_style(UiComponentStyles {
+                font_size: Some(CONTENT_FONT_SIZE),
+                ..Default::default()
+            })
+            .build()
+            .finish();
+
+        let max_label = appearance
+            .ui_builder()
+            .span(format!("{max}"))
+            .with_style(UiComponentStyles {
+                font_size: Some(CONTENT_FONT_SIZE),
+                ..Default::default()
+            })
+            .build()
+            .finish();
+
+        let current_value = AgentProfilesPageView::current_context_window_display_value(app)
+            .unwrap_or(cw.default_max)
+            .clamp(min, max);
+        let slider = appearance
+            .ui_builder()
+            .slider(view.context_window_slider_state.clone())
+            .with_range(min as f32..max as f32)
+            .with_default_value(current_value as f32)
+            .with_style(UiComponentStyles {
+                width: Some(CONTEXT_WINDOW_SLIDER_WIDTH),
+                margin: Some(Coords::default().left(8.).right(8.)),
+                ..Default::default()
+            })
+            .on_drag(|ctx, _, val| {
+                ctx.dispatch_typed_action(AgentProfilesPageAction::ContextWindowSliderDragged(
+                    val.round() as u32,
+                ));
+            })
+            .on_change(|ctx, _, val| {
+                ctx.dispatch_typed_action(AgentProfilesPageAction::SetContextWindowSize(
+                    val.round() as u32,
+                ));
+            })
+            .build()
+            .finish();
+
+        let context_window_editor = view.context_window_editor.clone();
+        let input_box = Dismiss::new(
+            appearance
+                .ui_builder()
+                .text_input(view.context_window_editor.clone())
+                .with_style(UiComponentStyles {
+                    width: Some(CONTEXT_WINDOW_INPUT_BOX_WIDTH),
+                    padding: Some(Coords {
+                        top: 6.,
+                        bottom: 6.,
+                        left: 10.,
+                        right: 10.,
+                    }),
+                    margin: Some(Coords::default().left(12.)),
+                    background: Some(appearance.theme().surface_2().into()),
+                    ..Default::default()
+                })
+                .build()
+                .finish(),
+        )
+        .on_dismiss(move |ctx, app| {
+            let buffer_text = context_window_editor.as_ref(app).buffer_text(app);
+            let cleaned: String = buffer_text
+                .chars()
+                .filter(|c| !c.is_whitespace() && *c != ',')
+                .collect();
+            if let Ok(parsed) = cleaned.parse::<u32>() {
+                ctx.dispatch_typed_action(AgentProfilesPageAction::SetContextWindowSize(parsed));
+            }
+        })
+        .finish();
+
+        let row = Flex::row()
+            .with_cross_axis_alignment(CrossAxisAlignment::Center)
+            .with_child(min_label)
+            .with_child(Shrinkable::new(1., slider).finish())
+            .with_child(max_label)
+            .with_child(input_box)
+            .finish();
+
+        let mut column = Flex::column().with_child(label).with_child(row);
+        if AgentProfilesPageView::active_profile_data(app)
+            .should_show_long_context_pricing_warning(view.dragged_context_window_value, app)
+        {
+            column.add_child(render_warning_box(
+                WarningBoxConfig::formatted_title(long_context_pricing_warning_title()),
+                appearance,
+            ));
+        }
+
+        Some(column.finish())
+    }
+
+    fn render_permissions_section(
+        &self,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let is_any_ai_enabled = ai_settings.is_any_ai_enabled(app);
+        let permissions_subheader = Container::new(render_custom_size_header(
+            appearance,
+            "Permissions",
+            14.0,
+            Some(styles::header_font_color(is_any_ai_enabled, app)),
+        ))
+        .with_margin_bottom(4.0)
+        .finish();
+
+        let code_diff_setting =
+            BlocklistAIPermissions::as_ref(app).get_apply_code_diffs_setting(app, None);
+        let code_diffs = self.render_execution_profile_dropdown(
+            "Apply code diffs",
+            Icon::Code2,
+            code_diff_setting.description(),
+            &view.apply_code_diffs_dropdown_menu,
+            ai_settings,
+            appearance,
+            app,
+        );
+
+        let read_files_setting =
+            BlocklistAIPermissions::as_ref(app).get_read_files_setting(app, None);
+        let mut read_files_flex = Flex::column().with_main_axis_size(MainAxisSize::Min);
+        read_files_flex.add_child(self.render_execution_profile_dropdown(
+            "Read files",
+            Icon::Notebook,
+            read_files_setting.description(),
+            &view.read_files_dropdown_menu,
+            ai_settings,
+            appearance,
+            app,
+        ));
+
+        if read_files_setting == ActionPermission::AlwaysAsk {
+            let directory_allowlist =
+                BlocklistAIPermissions::as_ref(app).get_read_files_allowlist(app, None);
+            read_files_flex.add_child(
+                Container::new(Self::render_directory_allowlist(
+                    directory_allowlist,
+                    view,
+                    ai_settings,
+                    appearance,
+                    app,
+                ))
+                .with_margin_bottom(HEADER_PADDING)
+                .finish(),
+            );
+        }
+        let read_files = read_files_flex.finish();
+
+        let execute_commands_setting =
+            BlocklistAIPermissions::as_ref(app).get_execute_commands_setting(app, None);
+        let mut execute_commands_flex = Flex::column().with_main_axis_size(MainAxisSize::Min);
+        execute_commands_flex.add_child(self.render_execution_profile_dropdown(
+            "Execute commands",
+            Icon::Terminal,
+            execute_commands_setting.description(),
+            &view.execute_commands_dropdown_menu,
+            ai_settings,
+            appearance,
+            app,
+        ));
+
+        if execute_commands_setting == ActionPermission::AlwaysAsk
+            || execute_commands_setting == ActionPermission::AgentDecides
+        {
+            let command_allowlist =
+                BlocklistAIPermissions::as_ref(app).get_execute_commands_allowlist(app, None);
+            execute_commands_flex.add_child(
+                Container::new(Self::render_command_allowlist(
+                    command_allowlist,
+                    view,
+                    ai_settings,
+                    appearance,
+                    app,
+                ))
+                .with_margin_bottom(8.)
+                .finish(),
+            );
+        }
+
+        if execute_commands_setting != ActionPermission::AlwaysAsk {
+            let command_denylist = Container::new(Self::render_command_denylist(
+                BlocklistAIPermissions::as_ref(app).get_execute_commands_denylist(app, None),
+                view,
+                ai_settings,
+                appearance,
+                app,
+            ))
+            .with_margin_bottom(8.)
+            .finish();
+            execute_commands_flex.add_child(command_denylist);
+        }
+        let execute_commands = execute_commands_flex.finish();
+
+        let mut widget_children = vec![permissions_subheader];
+
+        if UserWorkspaces::as_ref(app)
+            .ai_autonomy_settings()
+            .has_any_overrides()
+        {
+            widget_children.push(
+                Container::new(render_settings_info_banner(
+                    "Some of your permissions are managed by your workspace.",
+                    None,
+                    appearance,
+                ))
+                .with_margin_bottom(12.0)
+                .finish(),
+            );
+        }
+
+        widget_children.extend([code_diffs, read_files, execute_commands]);
+
+        let write_to_pty_setting =
+            BlocklistAIPermissions::as_ref(app).get_write_to_pty_setting(app, None);
+        let write_to_pty = self.render_execution_profile_dropdown(
+            "Interact with running commands",
+            Icon::Workflow,
+            write_to_pty_setting.description(),
+            &view.write_to_pty_autonomy_dropdown_menu,
+            ai_settings,
+            appearance,
+            app,
+        );
+        widget_children.push(write_to_pty);
+
+        if should_show_mcp_servers() {
+            let mcp_permissions = self.render_mcp_permissions(view, ai_settings, appearance, app);
+            widget_children.push(mcp_permissions);
+        }
+
+        if !FeatureFlag::FullSourceCodeEmbedding.is_enabled() {
+            let codebase_context = Self::render_codebase_context_outline_generation_setting(
+                self.codebase_context_toggle.clone(),
+                self.codebase_context_link_index.clone(),
+                view,
+                ai_settings,
+                appearance,
+                app,
+            );
+            widget_children.push(codebase_context);
+        }
+
+        Flex::column().with_children(widget_children).finish()
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    fn render_execution_profile_dropdown(
+        &self,
+        header_text: &str,
+        header_icon: Icon,
+        permission_description: &'static str,
+        dropdown_menu: &ViewHandle<Dropdown<AgentProfilesPageAction>>,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &warpui::AppContext,
+    ) -> Box<dyn Element> {
+        let header = Container::new(render_body_item_label_with_icon::<AgentProfilesPageAction>(
+            header_text.into(),
+            header_icon,
+            Some(styles::header_font_color(
+                ai_settings.is_any_ai_enabled(app),
+                app,
+            )),
+            None,
+            LocalOnlyIconState::Hidden,
+            ToggleState::Enabled,
+            appearance,
+        ))
+        .finish();
+
+        let description_color = appearance.theme().disabled_ui_text_color();
+        let alert_icon = Container::new(
+            ConstrainedBox::new(
+                Icon::AlertCircle
+                    .to_warpui_icon(
+                        appearance
+                            .theme()
+                            .sub_text_color(appearance.theme().surface_2()),
+                    )
+                    .finish(),
+            )
+            .with_width(14.)
+            .with_height(14.)
+            .finish(),
+        )
+        .with_margin_right(4.)
+        .finish();
+        let text = Text::new(
+            permission_description,
+            appearance.ui_font_family(),
+            appearance.ui_font_size(),
+        )
+        .with_color(description_color.into())
+        .finish();
+        let description = Flex::row()
+            .with_children([alert_icon, Shrinkable::new(1.0, text).finish()])
+            .finish();
+
+        Container::new(
+            Flex::column()
+                .with_child(header)
+                .with_child(ChildView::new(dropdown_menu).finish())
+                .with_child(description)
+                .finish(),
+        )
+        .with_margin_bottom(12.)
+        .finish()
+    }
+
+    fn render_command_denylist(
+        command_denylist: Vec<AgentModeCommandExecutionPredicate>,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let ai_disabled = !ai_settings.is_any_ai_enabled(app);
+        let org_denylist = BlocklistAIPermissions::get_org_execute_commands_denylist(app);
+        let mut tooltip_idx = 0usize;
+        let list = render_input_list(
+            None,
+            command_denylist
+                .into_iter()
+                .zip(view.command_denylist_mouse_state_handles.clone())
+                .rev()
+                .map(|(cmd, mouse_state_handle)| {
+                    let is_org = org_denylist.contains(&cmd);
+                    let tooltip_mouse_state = if is_org {
+                        let handle = view
+                            .command_denylist_tooltip_mouse_state_handles
+                            .get(tooltip_idx)
+                            .cloned();
+                        tooltip_idx += 1;
+                        handle
+                    } else {
+                        None
+                    };
+                    InputListItem {
+                        item: cmd.to_string(),
+                        mouse_state_handle,
+                        on_remove_action: AgentProfilesPageAction::RemoveFromProfileCommandDenylist(
+                            cmd,
+                        ),
+                        is_disabled: is_org || ai_disabled,
+                        tooltip_mouse_state,
+                    }
+                }),
+            Some(&view.command_denylist_editor),
+            appearance,
+        );
+        render_ai_list(
+            "Command denylist",
+            "Regular expressions to match commands that the Warp Agent should always ask permission to execute.",
+            list,
+            view,
+            ai_settings,
+            app,
+        )
+    }
+
+    fn render_command_allowlist(
+        command_allowlist: Vec<AgentModeCommandExecutionPredicate>,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let disabled = !ai_settings.is_command_allowlist_editable(app);
+        let list = render_input_list(
+            None,
+            command_allowlist
+                .into_iter()
+                .zip(view.command_allowlist_mouse_state_handles.clone())
+                .rev()
+                .map(move |(cmd, mouse_state_handle)| InputListItem {
+                    item: cmd.to_string(),
+                    mouse_state_handle,
+                    on_remove_action: AgentProfilesPageAction::RemoveFromProfileCommandAllowlist(
+                        cmd,
+                    ),
+                    is_disabled: disabled,
+                    tooltip_mouse_state: None,
+                }),
+            Some(&view.command_allowlist_editor),
+            appearance,
+        );
+
+        render_ai_list(
+            "Command allowlist",
+            "Regular expressions to match commands that can be automatically executed by the Warp Agent.",
+            list,
+            view,
+            ai_settings,
+            app,
+        )
+    }
+
+    fn render_directory_allowlist(
+        directory_allowlist: Vec<PathBuf>,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let disabled = !ai_settings.is_directory_allowlist_editable(app);
+        let list = render_input_list(
+            None,
+            directory_allowlist
+                .clone()
+                .into_iter()
+                .zip(view.directory_allowlist_mouse_state_handles.clone())
+                .rev()
+                .map(move |(path, mouse_state_handle)| InputListItem {
+                    item: path.display().to_string(),
+                    mouse_state_handle,
+                    on_remove_action: AgentProfilesPageAction::RemoveFromProfileDirectoryAllowlist(
+                        path,
+                    ),
+                    is_disabled: disabled,
+                    tooltip_mouse_state: None,
+                }),
+            Some(&view.directory_allowlist_editor),
+            appearance,
+        );
+
+        render_ai_list(
+            "Directory allowlist",
+            "Give the agent file access to certain directories.",
+            list,
+            view,
+            ai_settings,
+            app,
+        )
+    }
+
+    fn render_base_model_setting(
+        &self,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let theme = appearance.theme();
+        let show_in_prompt_checkbox = {
+            let is_checked = *SessionSettings::as_ref(app).show_model_selectors_in_prompt;
+
+            let mut checkbox = appearance
+                .ui_builder()
+                .checkbox(self.show_in_prompt_checkbox.clone(), None)
+                .check(is_checked);
+
+            if !ai_settings.is_any_ai_enabled(app) {
+                checkbox = checkbox.disabled();
+            }
+
+            Container::new(
+                Flex::row()
+                    .with_cross_axis_alignment(CrossAxisAlignment::Center)
+                    .with_children([
+                        checkbox
+                            .build()
+                            .on_click(move |ctx, _, _| {
+                                ctx.dispatch_typed_action(
+                                    AgentProfilesPageAction::ToggleShowBaseModelPickerInPrompt,
+                                );
+                            })
+                            .finish(),
+                        appearance
+                            .ui_builder()
+                            .span("Show model picker in prompt".to_string())
+                            .with_style(UiComponentStyles {
+                                font_color: Some(
+                                    theme.sub_text_color(theme.surface_2()).into_solid(),
+                                ),
+                                font_size: Some(12.0),
+                                ..Default::default()
+                            })
+                            .build()
+                            .finish(),
+                    ])
+                    .finish(),
+            )
+            .with_margin_top(-6.0)
+            .with_margin_left(-4.0)
+            .finish()
+        };
+
+        render_dropdown_item(
+            appearance,
+            "Base model",
+            Some(
+                "This model serves as the primary engine behind the Warp Agent. It powers most interactions and invokes other models for tasks like planning or code generation when necessary. Warp may automatically switch to alternate models based on model availability or for auxiliary tasks such as conversation summarization.",
+            ),
+            Some(show_in_prompt_checkbox),
+            LocalOnlyIconState::Hidden,
+            (!ai_settings.is_any_ai_enabled(app))
+                .then(|| appearance.theme().disabled_ui_text_color()),
+            &view.base_model_dropdown,
+        )
+    }
+
+    fn render_codebase_context_outline_generation_setting(
+        codebase_context_toggle: SwitchStateHandle,
+        codebase_context_link_index: HighlightedHyperlink,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &warpui::AppContext,
+    ) -> Box<dyn Element> {
+        let code_settings = CodeSettings::as_ref(app);
+        let toggle = render_ai_setting_toggle::<CodebaseContextEnabled>(
+            "Codebase Context",
+            AgentProfilesPageAction::ToggleCodebaseContext,
+            *code_settings.codebase_context_enabled,
+            ai_settings.is_any_ai_enabled(app),
+            codebase_context_toggle,
+            &view.local_only_icon_tooltip_states,
+            app,
+        );
+
+        let codebase_context_description = vec![
+            FormattedTextFragment::plain_text(
+                "Allow the Warp Agent to generate an outline of your codebase that can be used for context. No code is ever stored on our servers. ",
+            ),
+            FormattedTextFragment::hyperlink(
+                "Learn more",
+                "https://docs.warp.dev/agents/capabilities/codebase-context",
+            ),
+        ];
+        let description = Container::new(
+            FormattedTextElement::new(
+                FormattedText::new([FormattedTextLine::Line(codebase_context_description)]),
+                CONTENT_FONT_SIZE,
+                appearance.ui_font_family(),
+                appearance.ui_font_family(),
+                styles::description_font_color(ai_settings.is_any_ai_enabled(app), app).into(),
+                codebase_context_link_index,
+            )
+            .with_hyperlink_font_color(appearance.theme().accent().into_solid())
+            .register_default_click_handlers(|url, ctx, _| {
+                ctx.dispatch_typed_action(AgentProfilesPageAction::HyperlinkClick(url));
+            })
+            .finish(),
+        )
+        .with_margin_top(styles::DESCRIPTION_NEGATIVE_MARGIN_OFFSET)
+        .with_margin_bottom(styles::DESCRIPTION_MARGIN_BOTTOM)
+        .with_margin_right(styles::TOGGLE_WIDTH_MARGIN)
+        .finish();
+
+        Flex::column()
+            .with_child(toggle)
+            .with_child(description)
+            .finish()
+    }
+
+    fn render_mcp_permissions(
+        &self,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let all_runnable_mcp_servers =
+            TemplatableMCPServerManager::get_all_cloud_synced_mcp_servers(app);
+        if all_runnable_mcp_servers.is_empty() {
+            self.render_mcp_permissions_zero_state(ai_settings, appearance, app)
+        } else {
+            self.render_mcp_permissions_with_servers(view, ai_settings, appearance, app)
+        }
+    }
+
+    fn render_mcp_permissions_zero_state(
+        &self,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let header = Container::new(render_body_item_label_with_icon::<AgentProfilesPageAction>(
+            "Call MCP servers".into(),
+            Icon::Dataflow,
+            Some(styles::header_font_color(
+                ai_settings.is_any_ai_enabled(app),
+                app,
+            )),
+            None,
+            LocalOnlyIconState::Hidden,
+            ToggleState::Enabled,
+            appearance,
+        ))
+        .with_margin_bottom(4.)
+        .finish();
+
+        let subtext = {
+            let subtext_fragments = vec![
+                FormattedTextFragment::plain_text(
+                    "You haven't added any MCP servers yet. Once you do, you'll be able to control how much autonomy the Warp Agent has when interacting with them. ",
+                ),
+                FormattedTextFragment::hyperlink_action(
+                    "Add a server",
+                    AgentProfilesPageAction::OpenMCPServerCollection,
+                ),
+                FormattedTextFragment::plain_text(" or "),
+                FormattedTextFragment::hyperlink(
+                    "learn more about MCPs.",
+                    "https://docs.warp.dev/agents/capabilities/mcp",
+                ),
+            ];
+
+            Container::new(
+                FormattedTextElement::new(
+                    FormattedText::new([FormattedTextLine::Line(subtext_fragments)]),
+                    CONTENT_FONT_SIZE,
+                    appearance.ui_font_family(),
+                    appearance.ui_font_family(),
+                    styles::description_font_color(ai_settings.is_any_ai_enabled(app), app).into(),
+                    HighlightedHyperlink::default(),
+                )
+                .with_hyperlink_font_color(appearance.theme().accent().into_solid())
+                .register_default_click_handlers_with_action_support(|hyperlink_lens, ctx, _app| {
+                    match hyperlink_lens {
+                        HyperlinkLens::Url(url) => {
+                            ctx.dispatch_typed_action(AgentProfilesPageAction::HyperlinkClick(
+                                HyperlinkUrl {
+                                    url: url.to_owned(),
+                                },
+                            ));
+                        }
+                        HyperlinkLens::Action(action_ref) => {
+                            if let Some(action) = action_ref
+                                .as_any()
+                                .downcast_ref::<AgentProfilesPageAction>()
+                            {
+                                ctx.dispatch_typed_action(action.clone());
+                            }
+                        }
+                    }
+                })
+                .finish(),
+            )
+            .with_margin_bottom(4.0)
+            .finish()
+        };
+
+        Flex::column()
+            .with_child(header)
+            .with_child(subtext)
+            .finish()
+    }
+
+    fn render_mcp_permissions_with_servers(
+        &self,
+        view: &AgentProfilesPageView,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let mut column = Flex::column();
+
+        let current_mcp_setting =
+            BlocklistAIPermissions::as_ref(app).get_mcp_permissions_setting(app, None);
+
+        let permission_setting = self.render_execution_profile_dropdown(
+            "Call MCP servers",
+            Icon::Dataflow,
+            current_mcp_setting.description(),
+            &view.mcp_permissions_dropdown_menu,
+            ai_settings,
+            appearance,
+            app,
+        );
+        column.add_child(permission_setting);
+
+        if current_mcp_setting == ActionPermission::AlwaysAsk
+            || current_mcp_setting == ActionPermission::AgentDecides
+        {
+            let allowlist = self.render_mcp_list(
+                "MCP allowlist",
+                "Allow the Warp Agent to call these MCP servers.",
+                &view.mcp_allowlist_dropdown,
+                BlocklistAIPermissions::as_ref(app).get_mcp_allowlist(app, None),
+                view.mcp_allowlist_mouse_state_handles.clone(),
+                AgentProfilesPageAction::RemoveFromMCPAllowlist,
+                ai_settings,
+                appearance,
+                app,
+            );
+            column.add_child(allowlist);
+        }
+
+        if current_mcp_setting == ActionPermission::AlwaysAllow
+            || current_mcp_setting == ActionPermission::AgentDecides
+        {
+            let denylist = self.render_mcp_list(
+                "MCP denylist",
+                "The Warp Agent will always ask for permission before calling any MCP servers on this list.",
+                &view.mcp_denylist_dropdown,
+                BlocklistAIPermissions::as_ref(app).get_mcp_denylist(app, None),
+                view.mcp_denylist_mouse_state_handles.clone(),
+                AgentProfilesPageAction::RemoveFromMCPDenylist,
+                ai_settings,
+                appearance,
+                app,
+            );
+            column.add_child(denylist);
+        }
+
+        column.finish()
+    }
+
+    // Helper function to render the allow and denylists for mcp servers
+    #[allow(clippy::too_many_arguments)]
+    fn render_mcp_list(
+        &self,
+        title: &str,
+        description: &str,
+        dropdown: &ViewHandle<FilterableDropdown<AgentProfilesPageAction>>,
+        items: Vec<uuid::Uuid>,
+        mouse_state_handles: Vec<MouseStateHandle>,
+        action: impl Fn(uuid::Uuid) -> AgentProfilesPageAction,
+        ai_settings: &AISettings,
+        appearance: &Appearance,
+        app: &AppContext,
+    ) -> Box<dyn Element> {
+        let selector = Container::new(
+            Flex::row()
+                .with_main_axis_size(MainAxisSize::Max)
+                .with_main_axis_alignment(MainAxisAlignment::SpaceBetween)
+                .with_children(vec![
+                    Shrinkable::new(
+                        1.0,
+                        Container::new(render_dropdown_item_label(
+                            title.to_string(),
+                            Some(description.to_string()),
+                            LocalOnlyIconState::Hidden,
+                            (!ai_settings.is_any_ai_enabled(app))
+                                .then(|| appearance.theme().disabled_ui_text_color()),
+                            appearance,
+                        ))
+                        .with_margin_right(4.)
+                        .finish(),
+                    )
+                    .finish(),
+                    ChildView::new(dropdown).finish(),
+                ])
+                .finish(),
+        )
+        .with_margin_bottom(2.)
+        .finish();
+
+        let disabled = !ai_settings.is_any_ai_enabled(app);
+        let items = render_input_list(
+            None,
+            items
+                .into_iter()
+                .rev()
+                .zip(mouse_state_handles.clone())
+                .filter_map(move |(uuid, mouse_state_handle)| {
+                    let server_name = TemplatableMCPServerManager::get_mcp_name(&uuid, app);
+                    server_name.map(|server_name| InputListItem {
+                        item: server_name,
+                        mouse_state_handle,
+                        on_remove_action: action(uuid),
+                        is_disabled: disabled,
+                        tooltip_mouse_state: None,
+                    })
+                }),
+            None,
+            appearance,
+        );
+
+        Container::new(Flex::column().with_children(vec![selector, items]).finish())
+            .with_margin_bottom(8.)
+            .finish()
+    }
+}
