+++ app/src/settings_view/warp_agent_page_tests.rs
